% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinAnnotation.R
\name{add_RevCon}
\alias{add_RevCon}
\title{add REV and zz entries - used for testing}
\usage{
add_RevCon(stringsAll, pattern_decoys = "REV_", pattern_contaminants = "zz")
}
\arguments{
\item{stringsAll}{character vector of protein IDs}

\item{pattern_decoys}{prefix for decoy entries}

\item{pattern_contaminants}{prefix for contaminant entries}
}
\description{
add REV and zz entries - used for testing
}
