% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_IBAQ.R
\name{aggregate_data}
\alias{aggregate_data}
\title{dataset transform data}
\usage{
aggregate_data(lfqdata, agg_method = c("medpolish", "lmrob", "topN"), N = 3)
}
\arguments{
\item{lfqdata}{LFQData object}

\item{agg_method}{aggregation method}

\item{N}{number of top peptides for topN aggregation}
}
\description{
dataset transform data
}
\examples{

xx <- prolfqua::sim_lfq_data_peptide_config()
lfqdata <- prolfqua::LFQData$new(xx$data, xx$config)
aggregated <- aggregate_data(lfqdata, agg_method = "medpolish")
aggregated$response()
aggregated <- aggregate_data(lfqdata, agg_method = "lmrob")
aggregated$response()
aggregated <- aggregate_data(lfqdata, agg_method = "topN")
aggregated$response()

}
