% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq_to_ann_data.R
\name{anndata_from_LFQData}
\alias{anndata_from_LFQData}
\title{convert lfqdata to anndata}
\usage{
anndata_from_LFQData(lfqdata, pannot)
}
\arguments{
\item{lfqdata}{LFQData object}

\item{pannot}{ProteinAnnotation object}
}
\description{
convert lfqdata to anndata
}
\examples{
# example code
library(prolfqua)
library(prolfquapp)
lfqdata <- prolfqua::sim_lfq_data_2Factor_config()
lfqdata <- LFQData$new(lfqdata$data, lfqdata$config)
lfqdata$data$protein_Id <- add_RevCon(lfqdata$data$protein_Id)
pids <- grep("^zz|^REV", unique(lfqdata$data$protein_Id), value = TRUE, invert = TRUE)
addannot <- data.frame(
  protein_Id = pids,
  description = stringi::stri_rand_strings(length(pids), 13)
)

addannot <- addannot |> tidyr::separate(protein_Id, c("cleanID", NA), remove = FALSE)
pannot <- ProteinAnnotation$new(lfqdata,
  addannot,
  description = "description",
  cleaned_ids = "cleanID",
  pattern_contaminants = "^zz",
  pattern_decoys = "^REV"
)

#debug(anndata_from_LFQData)
anndata_from_LFQData(lfqdata, pannot)
#anndataR::write_h5ad(adata, path = "test.h5ad", mode = "w")
}
