% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinAnnotation.R
\name{build_protein_annot}
\alias{build_protein_annot}
\title{build Dataset protein annot, defaults are compatible with DIANN}
\usage{
build_protein_annot(
  lfqdata,
  msdata,
  idcol = c(protein_Id = "Protein.Group"),
  cleaned_protein_id = "Protein.Group.2",
  protein_description = "fasta.header",
  exp_nr_children = "nrPeptides",
  full_id = "fasta.id",
  more_columns = c("fasta.id"),
  pattern_contaminants = "^zz|^CON",
  pattern_decoys = "REV_"
)
}
\arguments{
\item{lfqdata}{LFQData}

\item{msdata}{data frame}

\item{idcol}{named vector mapping protein ID column}

\item{cleaned_protein_id}{column with cleaned protein ID}

\item{protein_description}{column with protein description}

\item{exp_nr_children}{column with number of peptides}

\item{full_id}{column with full protein ID}

\item{more_columns}{additional columns to include}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}
}
\description{
build Dataset protein annot, defaults are compatible with DIANN
}
\examples{
# example code

}
