% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_output}
\alias{capture_output}
\title{capture output of function to send it to log}
\usage{
capture_output(expr)
}
\arguments{
\item{expr}{expression to capture output from}
}
\description{
capture output of function to send it to log
}
