% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{check_annotation}
\alias{check_annotation}
\title{check if all required columns in annotation file are there.}
\usage{
check_annotation(annot, QC = FALSE)
}
\description{
check if all required columns in annotation file are there.
}
