% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_helpers.R
\name{column_to_rownames}
\alias{column_to_rownames}
\title{convert tibble to data.frame with rownames}
\usage{
column_to_rownames(.data, var = "rowname")
}
\arguments{
\item{.data}{a tibble or data.frame}

\item{var}{name of the column with new row.names}
}
\value{
a data.frame with rownames
}
\description{
convert tibble to data.frame with rownames
}
\examples{
ind <- tibble::tibble(a = 1:3, rowname = letters[1:3])
column_to_rownames(ind)
}
