% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_bat_files}
\alias{copy_bat_files}
\title{copy bat files to run the DEA app on windows}
\usage{
copy_bat_files(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy bat files to run the DEA app on windows
}
