% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_DEA_DIANN}
\alias{copy_DEA_DIANN}
\title{copy Markdown and runscript for DIANN diann-output.tsv}
\usage{
copy_DEA_DIANN(workdir = getwd(), run_script = FALSE)
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscript for DIANN diann-output.tsv
}
