% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_DEA_FragPipe_DDA}
\alias{copy_DEA_FragPipe_DDA}
\title{copy Markdown and runscript for 2 Grp Fragpipe data.}
\usage{
copy_DEA_FragPipe_DDA(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscript for 2 Grp Fragpipe data.
}
\keyword{internal}
