% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_DEA_Metabolomics_Files}
\alias{copy_DEA_Metabolomics_Files}
\title{copy Markdown and runscripts for DEA}
\usage{
copy_DEA_Metabolomics_Files(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscripts for DEA
}
