% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_DEA_R6_Files}
\alias{copy_DEA_R6_Files}
\title{copy Markdown templates for DEA (R6-based CMD_DEA_V2.R)}
\usage{
copy_DEA_R6_Files(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown templates for DEA (R6-based CMD_DEA_V2.R)
}
