% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_docker_script}
\alias{copy_docker_script}
\title{copy dockerfile to run the DEA app}
\usage{
copy_docker_script(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy dockerfile to run the DEA app
}
