% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_shell_script}
\alias{copy_shell_script}
\title{copy shellscript to run the DEA app}
\usage{
copy_shell_script(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy shellscript to run the DEA app
}
