% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{dataset_extract_contrasts}
\alias{dataset_extract_contrasts}
\title{extect contrasts from dataset}
\usage{
dataset_extract_contrasts(annot, GRP2)
}
\description{
extect contrasts from dataset
}
\examples{

file <- system.file("application/dataset_csv/dataset 25.csv", package = "prolfquapp")
res <- readr::read_csv(file)
GRP2 <- make_DEA_config()
GRP2 <- dataset_extract_contrasts(res,GRP2)
stopifnot(length(GRP2$pop$Contrasts) == 0)

file <- system.file("application/dataset_csv/dataset 26.csv", package = "prolfquapp")
res <- readr::read_csv(file)
GRP2 <- dataset_extract_contrasts(res,GRP2)
stopifnot(length(GRP2$pop$Contrasts) == 3)

}
