% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\name{dataset_get_functions}
\alias{dataset_get_functions}
\title{get functions for creating datasets}
\usage{
dataset_get_functions(preprocess_functions)
}
\description{
get functions for creating datasets
}
\examples{

# Get dataset functions for DIANN
preprocess_functions <- prolfquapp::prolfqua_preprocess_functions[["DIANN"]]
dataset_funcs <- dataset_get_functions(preprocess_functions)

# Use the functions
\dontrun{
files <- dataset_funcs$files_fn("path/to/data")
dataset <- dataset_funcs$dataset_fn(files, "output_file.csv")
}
}
