% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_set_factors}
\alias{dataset_set_factors}
\title{set factors and sample names columns}
\usage{
dataset_set_factors(atable, peptide, REPEATED = TRUE)
}
\description{
set factors and sample names columns
}
