% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetfactors_deprecated.R
\name{dataset_set_factors_deprecated}
\alias{dataset_set_factors_deprecated}
\title{set factors and sample names columns}
\usage{
dataset_set_factors_deprecated(atable, msdata, REPEATED = TRUE, SAINT = FALSE)
}
\description{
set factors and sample names columns
}
\examples{

dataset_csv_examples <- system.file("application/dataset_csv",package = "prolfquapp")
files <- dir(dataset_csv_examples, pattern = "*.csv")

res <- list()
for (i in seq_along(files)) {
  print(i)
  res[[i]] <- readr::read_csv(file.path(dataset_csv_examples, files[i]))
}

for (i in seq_along(res)) {
  atable <- prolfqua::AnalysisTableAnnotation$new()
  #atable$fileName = "channel"
  atable$hierarchy[["protein_Id"]] <- c("Protein")
  atable$hierarchy[["peptide_Id"]] <- c("Peptide")
  tmp <- prolfquapp::dataset_set_factors(atable, res[[i]] )
  cat(i, " : " , length(tmp$atable$factors), "factors : ", paste(tmp$atable$factors, collapse = "; "), "\n")
}

}
