% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_APB.R
\name{dataset_template_APB}
\alias{dataset_template_APB}
\title{create dataset template from an apb-export file}
\usage{
dataset_template_APB(files)
}
\arguments{
\item{files}{list with the \code{.h5ad} file in \code{data}}
}
\description{
create dataset template from an apb-export file
}
