% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_default.R
\name{dataset_template_BGS}
\alias{dataset_template_BGS}
\title{create templte dataset for BGS data}
\usage{
dataset_template_BGS(files)
}
\value{
data.frame
}
\description{
create templte dataset for BGS data
}
