% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{dataset_template_diann}
\alias{dataset_template_diann}
\title{create dataset template from DIANN outputs}
\usage{
dataset_template_diann(files)
}
\arguments{
\item{files}{list with data and fasta file paths}
}
\description{
create dataset template from DIANN outputs
}
