% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{dataset_template_MAXQUANT}
\alias{dataset_template_MAXQUANT}
\title{create template dataset for MAXQUANT data}
\usage{
dataset_template_MAXQUANT(files)
}
\arguments{
\item{files}{list with data and fasta file paths}
}
\description{
create template dataset for MAXQUANT data
}
\seealso{
Other MaxQuant: 
\code{\link{MaxQuant}},
\code{\link{tidyMQ_Evidence}()},
\code{\link{tidyMQ_Peptides}()},
\code{\link{tidyMQ_ProteinGroups}()}
}
\concept{MaxQuant}
