% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{dataset_template_MSSTATS}
\alias{dataset_template_MSSTATS}
\title{create dataset template from MSStats data.}
\usage{
dataset_template_MSSTATS(files)
}
\arguments{
\item{files}{list with data and fasta file paths}
}
\description{
create dataset template from MSStats data.
}
