% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEAnalyse.R
\name{DEAnalysePeptideToProtein}
\alias{DEAnalysePeptideToProtein}
\title{Differential expression analysis from peptide input to protein output}
\description{
Differential expression analysis from peptide input to protein output

Differential expression analysis from peptide input to protein output
}
\details{
Runs facades that consume peptide-level measurements but emit protein-level
contrasts. The input \code{LFQData} keeps peptide hierarchy columns, while its
active hierarchy depth points to the protein level.
}
\section{Super class}{
\code{\link[prolfquapp:DEAnalyse]{prolfquapp::DEAnalyse}} -> \code{DEAnalysePeptideToProtein}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAnalysePeptideToProtein-build_default}{\code{DEAnalysePeptideToProtein$build_default()}}
\item \href{#method-DEAnalysePeptideToProtein-clone}{\code{DEAnalysePeptideToProtein$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="prolfquapp" data-topic="DEAnalyse" data-id="build_facade"><a href='../../prolfquapp/html/DEAnalyse.html#method-DEAnalyse-build_facade'><code>prolfquapp::DEAnalyse$build_facade()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prolfquapp" data-topic="DEAnalyse" data-id="filter_contrasts"><a href='../../prolfquapp/html/DEAnalyse.html#method-DEAnalyse-filter_contrasts'><code>prolfquapp::DEAnalyse$filter_contrasts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prolfquapp" data-topic="DEAnalyse" data-id="get_annotated_contrasts"><a href='../../prolfquapp/html/DEAnalyse.html#method-DEAnalyse-get_annotated_contrasts'><code>prolfquapp::DEAnalyse$get_annotated_contrasts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prolfquapp" data-topic="DEAnalyse" data-id="initialize"><a href='../../prolfquapp/html/DEAnalyse.html#method-DEAnalyse-initialize'><code>prolfquapp::DEAnalyse$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalysePeptideToProtein-build_default"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalysePeptideToProtein-build_default}{}}}
\subsection{Method \code{build_default()}}{
Build the default peptide-to-protein facade.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalysePeptideToProtein$build_default()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalysePeptideToProtein-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalysePeptideToProtein-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalysePeptideToProtein$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
