% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{diann_output_to_peptide}
\alias{diann_output_to_peptide}
\title{Create peptide level (stripped sequences) report by aggregating Precursor abundances.}
\usage{
diann_output_to_peptide(report2)
}
\arguments{
\item{report2}{filtered DIA-NN report data frame}
}
\description{
\code{\link{diann_read_output}}
}
