% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{diann_read_output}
\alias{diann_read_output}
\title{read DiaNN diann-output.tsv file}
\usage{
diann_read_output(data, Lib.PG.Q.Value = 0.01, PG.Q.Value = 0.05)
}
\description{
filter for 2 peptides per protein, and for Q.Value < 0.01 (default)
}
\examples{

xx <- readr::read_tsv("inst/application/DIANN/2517219/out-2024-06-12/WU304486_report.tsv")
report2 <- prolfquapp::diann_read_output(xx)
report2$Protein.Group.2 <- sapply(report2$Protein.Group, function(x){ unlist(strsplit(x, "[ ;]"))[1]} )


}
