% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_deanalyse.R
\name{example_deanalyse}
\alias{example_deanalyse}
\title{Create an example DEAnalyse object from simulated data}
\usage{
example_deanalyse(Nprot = 100)
}
\arguments{
\item{Nprot}{number of simulated proteins (default 100)}
}
\value{
a \code{DEAnalyse} R6 object with contrasts computed and annotated
}
\description{
Builds a complete DEAnalyse R6 object using simulated peptide data.
Useful for vignette defaults, examples, and testing.
}
\examples{
dea <- example_deanalyse(Nprot = 10)
dea$contrast_results[[dea$default_model]]$get_contrasts()

}
