% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_deanalyse.R
\name{example_qc_generator}
\alias{example_qc_generator}
\title{Create an example QC_generator object from simulated data}
\usage{
example_qc_generator(Nprot = 100)
}
\arguments{
\item{Nprot}{number of simulated proteins (default 100)}
}
\value{
a \code{QC_generator} R6 object
}
\description{
Builds a QC_generator R6 object with simulated peptide data and
fake FASTA-derived columns (protein_length, nr_tryptic_peptides)
so that IBAQ computation works. Useful for vignette defaults and testing.
}
\examples{
pap <- example_qc_generator(Nprot = 10)
pap$get_prot_data()

}
