% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2GrpAnalysis.R
\name{exp2}
\alias{exp2}
\title{transform lfq data with x^2 - apply if non log data is needed}
\usage{
exp2(lfqTrans)
}
\arguments{
\item{lfqTrans}{transformed LFQData}
}
\description{
transform lfq data with x^2 - apply if non log data is needed
}
