% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{ExternalReader}
\alias{ExternalReader}
\title{project specification R6 class}
\description{
project specification R6 class

project specification R6 class
}
\examples{
ExternalReader$new()

}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{R6_extract_values}()},
\code{\link{make_DEA_config_R6}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{get_files}}{get_files}

\item{\code{get_files}}{preprocess}

\item{\code{extra_args}}{extra_args}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExternalReader-clone}{\code{ExternalReader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalReader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalReader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalReader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
