% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{extract_contrasts}
\alias{extract_contrasts}
\title{extract contrast from annotation file}
\usage{
extract_contrasts(annot, prefix = "G_", group = "group")
}
\description{
extract contrast from annotation file
}
