% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot_from_FASTA.R
\name{extract_GN}
\alias{extract_GN}
\title{extract gene names from uniprot 1sp fasta.headers}
\usage{
extract_GN(fasta.headers)
}
\arguments{
\item{fasta.headers}{vector with}
}
\description{
extract gene names from uniprot 1sp fasta.headers
}
