% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fasta_to_header}
\alias{fasta_to_header}
\title{extract fasta header from fasta file.}
\usage{
fasta_to_header(fasta)
}
\arguments{
\item{fasta}{\code{\link{prozor::readPeptideFasta}}}
}
\description{
extract fasta header from fasta file.
}
