% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{feature_annotation_get_best_score}
\alias{feature_annotation_get_best_score}
\title{get best feature annotation.}
\usage{
feature_annotation_get_best_score(x)
}
\arguments{
\item{x}{data frame of feature annotations}
}
\description{
get best feature annotation.
}
