% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{FragPipe}
\alias{FragPipe}
\title{Methods for reading Fragpipe outputs}
\description{
Convert FragPipe outputs into tidy tables. For more details see functions listed in the see also section.
}
\seealso{
Other FragPipe: 
\code{\link{tidy_FragPipe_combined_protein}()},
\code{\link{tidy_FragPipe_combined_protein_deprec}()}
}
\concept{FragPipe}
