% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R, R/generate_DEA_reports.R
\name{generate_DEA_reports}
\alias{generate_DEA_reports}
\title{Generate differential expression analysis reports}
\usage{
generate_DEA_reports(lfqdata, GRP2, prot_annot)

generate_DEA_reports(lfqdata, GRP2, prot_annot)
}
\arguments{
\item{prot_annot}{ProteinAnnotation$new}
}
\description{
Generate differential expression analysis reports

Generate differential expression analysis reports
}
