% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{generate_DEA_reports2}
\alias{generate_DEA_reports2}
\title{will replace generate_DEA_reports}
\usage{
generate_DEA_reports2(lfqdata, GRP2, prot_annot, Contrasts)
}
\arguments{
\item{lfqdata}{LFQData object}

\item{GRP2}{ProlfquAppConfig configuration object}

\item{prot_annot}{ProteinAnnotation object}

\item{Contrasts}{named character vector of contrasts}
}
\description{
will replace generate_DEA_reports
}
