% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot_from_FASTA.R
\name{get_annot_from_fasta}
\alias{get_annot_from_fasta}
\title{get_annot_from_fasta}
\usage{
get_annot_from_fasta(fasta.files, rev = "REV_", isUniprot = TRUE)
}
\description{
get_annot_from_fasta
}
