% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot_from_FASTA.R
\name{get_annot_from_fasta_V2}
\alias{get_annot_from_fasta_V2}
\title{get_annot_from_fasta}
\usage{
get_annot_from_fasta_V2(
  fasta.files,
  rev = "REV_",
  isUniprot = TRUE,
  min_length = 7,
  max_length = 30,
  include_seq = FALSE
)
}
\description{
get_annot_from_fasta
}
\examples{

#prolfquapp::get_annot_from_fasta(fasta.files)


}
