% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess_BGS_default.R
\name{get_BGS_files}
\alias{get_BGS_files}
\title{get BGS and fasta file location in folder}
\usage{
get_BGS_files(path, bgs_pattern = "*BGS Factory Report \\\\(Normal\\\\).tsv")
}
\value{
list with paths to data and fasta
}
\description{
get BGS and fasta file location in folder
}
\examples{
\dontrun{
x <- get_DIANN_files("inst/application/DIANN/2517219/")
}
}
