% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{get_CD_subset_columns}
\alias{get_CD_subset_columns}
\title{Detect subset columns in a CompoundDiscoverer long export}
\usage{
get_CD_subset_columns(long_file)
}
\arguments{
\item{long_file}{path to *_long.csv}
}
\value{
character vector of columns after Group
}
\description{
Detect subset columns in a CompoundDiscoverer long export
}
