% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{get_DIANN_files}
\alias{get_DIANN_files}
\title{get report.tsv and fasta file location in folder}
\usage{
get_DIANN_files(path)
}
\value{
list with paths to data and fasta
}
\description{
get report.tsv and fasta file location in folder
}
\examples{
\dontrun{
x <- get_DIANN_files("inst/application/DIANN/2517219/")
}
}
