% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{get_formula}
\alias{get_formula}
\title{will generates formula}
\usage{
get_formula(config, interaction = FALSE)
}
\arguments{
\item{config}{AnalysisConfiguration object}

\item{interaction}{if TRUE include interaction terms}
}
\description{
will generates formula
}
