% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_other_outputs.R
\name{get_FP_combined_STY_files}
\alias{get_FP_combined_STY_files}
\title{get report.tsv and fasta file location in folder}
\usage{
get_FP_combined_STY_files(path)
}
\value{
list with paths to data and fasta
}
\description{
get report.tsv and fasta file location in folder
}
