% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{get_MSstats_files}
\alias{get_MSstats_files}
\title{get petpide.txt and fasta file location in folder}
\usage{
get_MSstats_files(path)
}
\value{
list with paths to data and fasta
}
\description{
get petpide.txt and fasta file location in folder
}
