% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{get_mzMine_files}
\alias{get_mzMine_files}
\title{get mzmine fliles}
\usage{
get_mzMine_files(path)
}
\arguments{
\item{path}{path to data directory}
}
\description{
get mzmine fliles
}
\examples{
path <- "WU323671_mzMine_o35537_WpH9V2_neg_v2_result"
files <- get_mzMine_files(path)
}
