% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_lfqdata.R
\name{LFQData_from_anndata}
\alias{LFQData_from_anndata}
\title{Convert AnnData back to LFQData + ProteinAnnotation}
\usage{
LFQData_from_anndata(adata)
}
\arguments{
\item{adata}{an \code{anndataR::AnnData} object with \code{uns$prolfquapp}}
}
\value{
list with \code{lfqdata} (LFQData) and \code{protein_annotation}
  (ProteinAnnotation)
}
\description{
Reconstructs the \code{list(lfqdata, protein_annotation)} pair from
an AnnData object that was created by \code{\link{preprocess_DIANN_anndata}}
(or any producer that writes the \code{prolfquapp} uns namespace).
}
\examples{
\dontrun{
res <- sim_data_protAnnot()
adata <- preprocess_DIANN_anndata_from_lfq(res$lfqdata, res$pannot)
back <- LFQData_from_anndata(adata)
}
}
