% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinAnnotation.R
\name{make_annotated_experiment}
\alias{make_annotated_experiment}
\title{make lfqdata with row annotation}
\usage{
make_annotated_experiment(Nprot = 100)
}
\arguments{
\item{Nprot}{number of proteins to simulate}
}
\description{
make lfqdata with row annotation
}
