% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{make_CD_duplicate_features_unique}
\alias{make_CD_duplicate_features_unique}
\title{Make duplicated CompoundDiscoverer feature/sample rows unique}
\usage{
make_CD_duplicate_features_unique(
  data,
  feature_col = "Feature_ID",
  sample_col = "Sample"
)
}
\arguments{
\item{data}{long CompoundDiscoverer data frame}

\item{feature_col}{feature identifier column}

\item{sample_col}{sample identifier column}
}
\value{
data frame with unique feature identifiers for affected rows
}
\description{
Make duplicated CompoundDiscoverer feature/sample rows unique
}
