% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{make_DEA_config}
\alias{make_DEA_config}
\title{create GRP2 configuration.
Use this function if there is no Yaml Input.}
\usage{
make_DEA_config(
  ZIPDIR = ".",
  PROJECTID = "",
  ORDERID = "",
  WORKUNITID = "",
  Normalization = c("none", "vsn", "quantile", "robscale"),
  aggregation = c("medpolish", "top3", "lmrob"),
  Diffthreshold = 1,
  FDRthreshold = 0.1,
  removeContaminants = FALSE,
  removeDecoys = FALSE,
  patternDecoys = "^REV_",
  patternContaminants = "^zz",
  application = "FragPipeTMT",
  nrPeptides = 2
)
}
\arguments{
\item{ZIPDIR}{output zip directory}

\item{PROJECTID}{project identifier}

\item{ORDERID}{order identifier}

\item{WORKUNITID}{workunit identifier}

\item{Normalization}{normalization method}

\item{aggregation}{aggregation method}

\item{Diffthreshold}{fold-change difference threshold}

\item{FDRthreshold}{FDR significance threshold}

\item{removeContaminants}{if TRUE remove contaminants}

\item{removeDecoys}{if TRUE remove decoy sequences}

\item{patternDecoys}{default "^REV_"}

\item{patternContaminants}{default "^zz_"}

\item{application}{software application name}

\item{nrPeptides}{minimum number of peptides per protein}
}
\description{
create GRP2 configuration.
Use this function if there is no Yaml Input.
}
\examples{
DEAconfig <- make_DEA_config()
DEAconfig
}
