% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{make_DEA_config_R6}
\alias{make_DEA_config_R6}
\title{create GRP2 configuration for differential expression analysis
Use this function if there is no Yaml Input.}
\usage{
make_DEA_config_R6(
  PATH = ".",
  PROJECTID = "",
  ORDERID = "",
  WORKUNITID = "",
  Normalization = c("none", "vsn", "robscale"),
  aggregation = c("medpolish", "topN", "rlm"),
  diff_threshold = 1,
  FDR_threshold = 0.1,
  nr_peptides = 1,
  removeContaminants = FALSE,
  removeDecoys = FALSE,
  patternDecoys = "^REV_|^rev_",
  patternContaminants = "^zz|^CON|Cont_",
  application = "DIANN",
  prefix = "DEA",
  model = "lm_impute"
)
}
\arguments{
\item{PATH}{working directory path}

\item{PROJECTID}{project identifier}

\item{ORDERID}{order identifier}

\item{WORKUNITID}{workunit identifier}

\item{Normalization}{normalization method: "none", "vsn", "robscale"}

\item{aggregation}{aggregation method: "medpolish", "topN", "rlm"}

\item{diff_threshold}{difference threshold}

\item{FDR_threshold}{FDR threshold}

\item{nr_peptides}{number of peptides required}

\item{removeContaminants}{should contaminants be removed}

\item{removeDecoys}{should decoys be removed}

\item{patternDecoys}{pattern for decoy proteins}

\item{patternContaminants}{pattern for contaminant proteins}

\item{application}{software application name}

\item{prefix}{analysis prefix (DEA or QC)}

\item{model}{facade registry key for contrast analysis. Use "saint" for
SAINTexpress interaction scoring.}
}
\value{
ProlfquAppConfig R6 object
}
\description{
create GRP2 configuration for differential expression analysis
Use this function if there is no Yaml Input.
}
\examples{

DEAconfig <- make_DEA_config_R6(ORDERID = "1234", WORKUNITID = "1234")
DEAconfig$set_zipdir_name()
DEAconfig$get_zipdir()
DEAconfig$get_result_dir()
DEAconfig$get_input_dir()
R6list <- prolfqua::R6_extract_values(DEAconfig)

}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ExternalReader}},
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
