% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{make_DEA_report}
\alias{make_DEA_report}
\title{Create DEA report in html and write data to xlsx table}
\usage{
make_DEA_report(lfqdata, protAnnot, GRP2)
}
\arguments{
\item{lfqdata}{LFQData object}

\item{protAnnot}{ProteinAnnotation object}

\item{GRP2}{list with named arguments i.e. Contrasts, projectID, projectName, workunitID, nrPeptides, Diffthreshold, FDRthreshold}
}
\description{
For use examples see run_scripts directory
}
\seealso{
Other workflow: 
\code{\link{make_SummarizedExperiment}()}
}
\concept{workflow}
