% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{make_feature_annotation}
\alias{make_feature_annotation}
\title{get feature annotation.}
\usage{
make_feature_annotation(x, .loader = feature_annotation_get_best_score)
}
\arguments{
\item{x}{data frame of feature annotations}

\item{.loader}{function to select best annotations}
}
\description{
get feature annotation.
}
\examples{
if (FALSE) {
x <- readr::read_csv("outputs-20250407T1707/mzmine/result_annotations.csv")
bestscore <- make_feature_annotation(x)
}
}
