% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{make_SummarizedExperiment}
\alias{make_SummarizedExperiment}
\title{Convert prolfqua differential expression analysis results to SummarizedExperiment}
\usage{
make_SummarizedExperiment(
  GRP2,
  colname = NULL,
  rowname = NULL,
  strip = "~lfq~light",
  .url_builder = bfabric_url_builder
)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}
}
\value{
SummarizedExperiment
}
\description{
Convert prolfqua differential expression analysis results to SummarizedExperiment
}
\seealso{
Other workflow: 
\code{\link{make_DEA_report}()},
\code{\link{write_DEA}()}
}
\concept{workflow}
