% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{massage_CD}
\alias{massage_CD}
\title{massage CD output compound table.}
\usage{
massage_CD(in_file, EXCEL = TRUE)
}
\arguments{
\item{in_file}{path to Compound Discoverer file or data frame}

\item{EXCEL}{if TRUE read Excel format}
}
\description{
massage CD output compound table.
}
