% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_MaxQuant.R
\name{MaxQuant}
\alias{MaxQuant}
\title{Methods for reading  MaxQuant outputs}
\description{
Convert MaxQuant outputs into tidy tables. For more details see functions listed in the see also section.
}
\seealso{
Other MaxQuant: 
\code{\link{tidyMQ_Evidence}()},
\code{\link{tidyMQ_Peptides}()},
\code{\link{tidyMQ_ProteinGroups}()}
}
\concept{MaxQuant}
