% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_path}
\alias{normalize_path}
\title{Function to normalize paths for both Windows and Linux}
\usage{
normalize_path(paths, os = .Platform$OS.type)
}
\arguments{
\item{paths}{character vector of file paths}

\item{os}{operating system type}
}
\value{
normalized path
}
\description{
Function to normalize paths for both Windows and Linux
}
\examples{
exp_paths <-c("E:\\\\projects\\\\p29033\\\\TKOiWAT\\\\20240123_015_S629149_iWAT_FL1.d",
  "E:\\\\projects\\\\p29033\\\\TKOiWAT\\\\20240123_016_S629150_iWAT_FL2.d",
  "E:\\\\projects\\\\p29033\\\\TKOiWAT\\\\20240123_017_S629151_iWAT_FL3.d")

normalize_path(exp_paths)

}
