% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot_from_FASTA.R
\name{nr_tryptic_peptides}
\alias{nr_tryptic_peptides}
\title{Compute number of tryptic peptides}
\usage{
nr_tryptic_peptides(sequence, min_length = 6, max_length = 30)
}
\arguments{
\item{sequence}{amino acid sequence}

\item{min_length}{minimum peptide length}

\item{max_length}{maximum peptide length}
}
\description{
Compute number of tryptic peptides
}
\examples{
# example code

sequence <- "MKGLPRAKSHGSTGWGKRKRNKPK"
nr_tryptic_peptides(sequence, min_length=5)

}
