% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_default.R,
%   R/proprocess_BGS_default.R
\name{preprocess_BGS}
\alias{preprocess_BGS}
\title{preprocess DIANN ouput, filter by q_value and nr_peptides}
\usage{
preprocess_BGS(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev",
  q_value = 0.01,
  hierarchy_depth = 2
)

preprocess_BGS(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev",
  q_value = 0.01,
  hierarchy_depth = 2
)
}
\arguments{
\item{quant_data}{path to quantification data file}

\item{fasta_file}{path to fasta file(s)}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{q_value}{q-value threshold for filtering}

\item{hierarchy_depth}{hierarchy depth for aggregation}
}
\value{
list with lfqdata and protein annotation

list with lfqdata and protein annotation
}
\description{
preprocess DIANN ouput, filter by q_value and nr_peptides

preprocess DIANN ouput, filter by q_value and nr_peptides
}
\examples{
\dontrun{
x <- get_BGS_files("DefaultParsing")
bgs <- read_BGS(x$data)
annot <- data.frame(raw.file = bgs$R.FileName |> unique(),
 Name = paste(c(rep("A",3),rep("B",3)),1:6, sep="_"),
group = c(rep("A",3),rep("B",3)))
annotation <- annot |> prolfquapp::read_annotation(QC = TRUE)
#debug(preprocess_BGS)
xd <- preprocess_BGS(x$data, x$fasta, annotation)
}
\dontrun{
x <- get_BGS_files("DefaultParsing")
bgs <- read_BGS(x$data)
annot <- data.frame(raw.file = bgs$R.FileName |> unique(),
 Name = paste(c(rep("A",3),rep("B",3)),1:6, sep="_"),
group = c(rep("A",3),rep("B",3)))
annotation <- annot |> prolfquapp::read_annotation(QC = TRUE)
#debug(preprocess_BGS)
xd <- preprocess_BGS(x$data, x$fasta, annotation)
}
}
