% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{preprocess_CD}
\alias{preprocess_CD}
\title{load compound discoverer (CD) files}
\usage{
preprocess_CD(xdl, annotation)
}
\arguments{
\item{xdl}{data frame from massage_CD}

\item{annotation}{list returned by `read_annotation` function}
}
\value{
A list containing the prepared \code{LFQData} and
  \code{ProteinAnnotation} objects.
}
\description{
load compound discoverer (CD) files
}
