% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN_anndata.R
\name{preprocess_DIANN_anndata}
\alias{preprocess_DIANN_anndata}
\title{Preprocess DIANN output and return AnnData}
\usage{
preprocess_DIANN_anndata(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev",
  q_value = 0.01,
  hierarchy_depth = 1,
  nr_peptides = 1
)
}
\arguments{
\item{quant_data}{path to quantification data file}

\item{fasta_file}{path to fasta file(s)}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{q_value}{q-value threshold for filtering}

\item{hierarchy_depth}{hierarchy depth for aggregation}

\item{nr_peptides}{minimum number of peptides per protein}
}
\value{
\code{anndataR::AnnData} object
}
\description{
Same interface as \code{\link{preprocess_DIANN}} but returns an
\code{anndataR::AnnData} object instead of
\code{list(lfqdata, protein_annotation)}.
}
\details{
The AnnData \code{uns} slot contains three namespaces:
\describe{
  \item{X_layer_name}{Name of the primary intensity column stored in X}
  \item{exploreDE}{Column role metadata compatible with anndata_omics_bridge}
  \item{prolfquapp}{Round-trip reconstruction metadata (config, protein annotation)}
}
}
\examples{
\dontrun{
x <- get_DIANN_files("inst/application/DIANN/2706527/")
annotation <- file.path("inst/application/DIANN/2706527/dataset.csv") |>
  readr::read_csv() |>
  prolfquapp::read_annotation(QC = TRUE)
adata <- preprocess_DIANN_anndata(x$data, x$fasta, annotation)
}
}
