% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dummy.R
\name{preprocess_dummy}
\alias{preprocess_dummy}
\title{preprocess_dummy}
\usage{
preprocess_dummy(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev"
)
}
\arguments{
\item{quant_data}{path to quantification data file}

\item{fasta_file}{path to fasta file(s)}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}
}
\description{
preprocess_dummy
}
