% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_other_outputs.R
\name{preprocess_FP_multi_site}
\alias{preprocess_FP_multi_site}
\title{preprocess FP multisite, filter by purity_threshold and PeptideProphetProb}
\usage{
preprocess_FP_multi_site(
  quant_data,
  fasta,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess FP multisite, filter by purity_threshold and PeptideProphetProb
}
