% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{preprocess_FP_multisite}
\alias{preprocess_FP_multisite}
\title{preprocess FP multisite, filter by purity_threshold and PeptideProphetProb}
\usage{
preprocess_FP_multisite(quant_data, fasta, annotation, fasta.rev = "REV_")
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess FP multisite, filter by purity_threshold and PeptideProphetProb
}
