% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{preprocess_FP_PSM}
\alias{preprocess_FP_PSM}
\title{preprocess DIANN ouput, filter by q_value and nr_peptides}
\usage{
preprocess_FP_PSM(
  quant_data,
  fasta_file,
  annotation,
  q_value = 0.01,
  nrPeptides = 1
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess DIANN ouput, filter by q_value and nr_peptides
}
