% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{preprocess_MSstats}
\alias{preprocess_MSstats}
\title{preprocess MQ peptide}
\usage{
preprocess_MSstats(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON",
  pattern_decoys = "REV_"
)
}
\description{
preprocess MQ peptide
}
