% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{preprocess_mzMine}
\alias{preprocess_mzMine}
\title{preprocess mzMine input}
\usage{
preprocess_mzMine(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = NULL,
  pattern_decoys = NULL,
  annotated = FALSE,
  nr_peptides = 1
)
}
\arguments{
\item{quant_data}{path to mzMine features csv file}

\item{fasta_file}{path to annotations csv file}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{annotated}{if TRUE only keep annotated features}

\item{nr_peptides}{accepted for interface uniformity but ignored (mzMine features have one child per protein)}
}
\value{
A list containing the prepared \code{LFQData} and
  \code{ProteinAnnotation} objects.
}
\description{
preprocess mzMine input
}
\examples{
if(FALSE){
annotation <- read_annotation(readr::read_tsv("outputs-20250407T1707/bfabric/input_dataset.tsv"), QC = TRUE)
files <- get_mzMine_files("outputs-20250407T1707/")
res <- preprocess_mzMine(files$data, files$fasta , annotation, annotated = TRUE)
}
}
