% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\name{preprocess_software_deprec}
\alias{preprocess_software_deprec}
\title{collects preprocess methods for various software}
\usage{
preprocess_software_deprec(
  indir,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^rev_|^REV_",
  software = c("DIANN", "FP_TMT", "MAXQUANT", "MSSTATS", "MSSTATS_FP_DIA")
)
}
\description{
collects preprocess methods for various software
}
