% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppConfigurationR6.R
\name{ProjectSpec}
\alias{ProjectSpec}
\title{project specification R6 class}
\description{
project specification R6 class

project specification R6 class
}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ProcessingOptions}},
\code{\link{ProlfquAppConfig}},
\code{\link{R6_extract_values}()},
\code{\link{make_DEA_config_R6}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{project_Id}}{project ID}

\item{\code{project_name}}{project name}

\item{\code{order_Id}}{order ID}

\item{\code{workunit_Id}}{workunit ID}

\item{\code{input_URL}}{input URL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectSpec-clone}{\code{ProjectSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
