% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Configuration.R
\name{ProlfquAppConfig}
\alias{ProlfquAppConfig}
\title{R6 class representing}
\description{
R6 class representing

R6 class representing
}
\examples{


r6obj_config <- ProlfquAppConfig$new(ProcessingOptions$new(), ProjectSpec$new())
xx <- R6_extract_values(r6obj_config)
yaml::write_yaml(xx, file = "test.yaml")
config <- yaml::read_yaml(file = "test.yaml")
}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{R6_extract_values}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{processing_options}}{ProcessingOption R6 class}

\item{\code{project_spec}}{Project Spec R6 class}

\item{\code{software}}{name of input software}

\item{\code{zipdir}}{results should go to zipdir}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProlfquAppConfig-new}{\code{ProlfquAppConfig$new()}}
\item \href{#method-ProlfquAppConfig-clone}{\code{ProlfquAppConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProlfquAppConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProlfquAppConfig-new}{}}}
\subsection{Method \code{new()}}{
set procession options and project spec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProlfquAppConfig$new(
  processing_options,
  project_spec,
  zipdir = NULL,
  software = "DIANN"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processing_options}}{instance of ProjectOptions}

\item{\code{project_spec}}{instance of ProjectSpec}

\item{\code{zipdir}}{where to store results}

\item{\code{software}}{name of input software}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProlfquAppConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProlfquAppConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProlfquAppConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
