% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_AbundancesR6.R
\name{ProteinAbundanceProcessor}
\alias{ProteinAbundanceProcessor}
\title{ProteinAbundanceProcessor}
\description{
ProteinAbundanceProcessor

ProteinAbundanceProcessor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProteinAbundanceProcessor-new}{\code{ProteinAbundanceProcessor$new()}}
\item \href{#method-ProteinAbundanceProcessor-get_peptides_wide}{\code{ProteinAbundanceProcessor$get_peptides_wide()}}
\item \href{#method-ProteinAbundanceProcessor-get_annotation}{\code{ProteinAbundanceProcessor$get_annotation()}}
\item \href{#method-ProteinAbundanceProcessor-get_prot_data}{\code{ProteinAbundanceProcessor$get_prot_data()}}
\item \href{#method-ProteinAbundanceProcessor-get_prot_wide}{\code{ProteinAbundanceProcessor$get_prot_wide()}}
\item \href{#method-ProteinAbundanceProcessor-get_prot_IBAQ}{\code{ProteinAbundanceProcessor$get_prot_IBAQ()}}
\item \href{#method-ProteinAbundanceProcessor-get_protein_per_group_abundance}{\code{ProteinAbundanceProcessor$get_protein_per_group_abundance()}}
\item \href{#method-ProteinAbundanceProcessor-get_protein_per_group_abundance_with_row_annot}{\code{ProteinAbundanceProcessor$get_protein_per_group_abundance_with_row_annot()}}
\item \href{#method-ProteinAbundanceProcessor-get_protein_per_group_abundance_wide}{\code{ProteinAbundanceProcessor$get_protein_per_group_abundance_wide()}}
\item \href{#method-ProteinAbundanceProcessor-get_prot_IBAQ_wide}{\code{ProteinAbundanceProcessor$get_prot_IBAQ_wide()}}
\item \href{#method-ProteinAbundanceProcessor-get_list}{\code{ProteinAbundanceProcessor$get_list()}}
\item \href{#method-ProteinAbundanceProcessor-write_xlsx}{\code{ProteinAbundanceProcessor$write_xlsx()}}
\item \href{#method-ProteinAbundanceProcessor-render_QC_protein_abundances}{\code{ProteinAbundanceProcessor$render_QC_protein_abundances()}}
\item \href{#method-ProteinAbundanceProcessor-render_sample_size_QC}{\code{ProteinAbundanceProcessor$render_sample_size_QC()}}
\item \href{#method-ProteinAbundanceProcessor-get_protein_per_group_small_wide}{\code{ProteinAbundanceProcessor$get_protein_per_group_small_wide()}}
\item \href{#method-ProteinAbundanceProcessor-clone}{\code{ProteinAbundanceProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$new(lfqdata, protein_annotation, prolfquapp_config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_peptides_wide"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_peptides_wide}{}}}
\subsection{Method \code{get_peptides_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_peptides_wide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_annotation}{}}}
\subsection{Method \code{get_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_annotation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_prot_data"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_prot_data}{}}}
\subsection{Method \code{get_prot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_prot_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_prot_wide"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_prot_wide}{}}}
\subsection{Method \code{get_prot_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_prot_wide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_prot_IBAQ"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_prot_IBAQ}{}}}
\subsection{Method \code{get_prot_IBAQ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_prot_IBAQ()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_protein_per_group_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_protein_per_group_abundance}{}}}
\subsection{Method \code{get_protein_per_group_abundance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_protein_per_group_abundance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_protein_per_group_abundance_with_row_annot"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_protein_per_group_abundance_with_row_annot}{}}}
\subsection{Method \code{get_protein_per_group_abundance_with_row_annot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_protein_per_group_abundance_with_row_annot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_protein_per_group_abundance_wide"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_protein_per_group_abundance_wide}{}}}
\subsection{Method \code{get_protein_per_group_abundance_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_protein_per_group_abundance_wide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_prot_IBAQ_wide"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_prot_IBAQ_wide}{}}}
\subsection{Method \code{get_prot_IBAQ_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_prot_IBAQ_wide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_list}{}}}
\subsection{Method \code{get_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-write_xlsx"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-write_xlsx}{}}}
\subsection{Method \code{write_xlsx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$write_xlsx()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-render_QC_protein_abundances"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-render_QC_protein_abundances}{}}}
\subsection{Method \code{render_QC_protein_abundances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$render_QC_protein_abundances()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-render_sample_size_QC"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-render_sample_size_QC}{}}}
\subsection{Method \code{render_sample_size_QC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$render_sample_size_QC()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-get_protein_per_group_small_wide"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-get_protein_per_group_small_wide}{}}}
\subsection{Method \code{get_protein_per_group_small_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$get_protein_per_group_small_wide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinAbundanceProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinAbundanceProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinAbundanceProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
