% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinDataPrep.R
\name{ProteinDataPrep}
\alias{ProteinDataPrep}
\title{ProteinDataPrep}
\description{
ProteinDataPrep

ProteinDataPrep
}
\details{
Handles data preparation for differential expression analysis:
contaminant/decoy filtering, peptide-to-protein aggregation, and normalization.
}
\examples{
pep <- prolfqua::sim_lfq_data_peptide_config(Nprot = 100)
pep <- prolfqua::LFQData$new(pep$data, pep$config)
pA <- data.frame(protein_Id = unique(pep$data$protein_Id))
pA <- pA |> dplyr::mutate(fasta.annot = paste0(pA$protein_Id, "_description"))
pA <- prolfquapp::ProteinAnnotation$new(pep, row_annot = pA, description = "fasta.annot")
GRP2 <- prolfquapp::make_DEA_config_R6()
GRP2$processing_options$transform <- "robscale"

data_prep <- prolfquapp::ProteinDataPrep$new(pep, pA, GRP2)
data_prep$cont_decoy_summary()
data_prep$remove_cont_decoy()
data_prep$aggregate()
data_prep$transform_data()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{lfq_data_peptide}}{LFQData peptide level}

\item{\code{lfq_data}}{LFQData protein level (after aggregation)}

\item{\code{lfq_data_transformed}}{normalized LFQData}

\item{\code{lfq_data_peptide_transformed}}{transformed peptide-level LFQData (for nested facades)}

\item{\code{aggregator}}{aggregator object}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{summary}}{data.frame with contaminant/decoy summary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProteinDataPrep-new}{\code{ProteinDataPrep$new()}}
\item \href{#method-ProteinDataPrep-cont_decoy_summary}{\code{ProteinDataPrep$cont_decoy_summary()}}
\item \href{#method-ProteinDataPrep-remove_cont_decoy}{\code{ProteinDataPrep$remove_cont_decoy()}}
\item \href{#method-ProteinDataPrep-aggregate}{\code{ProteinDataPrep$aggregate()}}
\item \href{#method-ProteinDataPrep-get_aggregation_plots}{\code{ProteinDataPrep$get_aggregation_plots()}}
\item \href{#method-ProteinDataPrep-write_aggregation_plots}{\code{ProteinDataPrep$write_aggregation_plots()}}
\item \href{#method-ProteinDataPrep-transform_data}{\code{ProteinDataPrep$transform_data()}}
\item \href{#method-ProteinDataPrep-transform_peptide_data}{\code{ProteinDataPrep$transform_peptide_data()}}
\item \href{#method-ProteinDataPrep-build_deanalyse}{\code{ProteinDataPrep$build_deanalyse()}}
\item \href{#method-ProteinDataPrep-clone}{\code{ProteinDataPrep$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-new}{}}}
\subsection{Method \code{new()}}{
Initialize ProteinDataPrep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$new(lfq_data_peptide, rowAnnot, prolfq_app_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lfq_data_peptide}}{LFQData object at peptide level}

\item{\code{rowAnnot}}{ProteinAnnotation object}

\item{\code{prolfq_app_config}}{ProlfquAppConfig object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-cont_decoy_summary"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-cont_decoy_summary}{}}}
\subsection{Method \code{cont_decoy_summary()}}{
Count number of contaminants and decoys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$cont_decoy_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-remove_cont_decoy"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-remove_cont_decoy}{}}}
\subsection{Method \code{remove_cont_decoy()}}{
Remove contaminants and decoys from peptide data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$remove_cont_decoy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Aggregate peptide data to protein level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$aggregate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-get_aggregation_plots"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-get_aggregation_plots}{}}}
\subsection{Method \code{get_aggregation_plots()}}{
Get aggregation plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$get_aggregation_plots(exp_nr_children = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp_nr_children}}{minimum number of peptides per protein; default 2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-write_aggregation_plots"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-write_aggregation_plots}{}}}
\subsection{Method \code{write_aggregation_plots()}}{
Write aggregation plots to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$write_aggregation_plots(exp_nr_children = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp_nr_children}}{minimum number of peptides per protein; default 2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
Transform and normalize protein-level data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-transform_peptide_data"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-transform_peptide_data}{}}}
\subsection{Method \code{transform_peptide_data()}}{
Transform peptide-level data (for nested facades like lmer/ropeca)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$transform_peptide_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-build_deanalyse"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-build_deanalyse}{}}}
\subsection{Method \code{build_deanalyse()}}{
Build a DEAnalyse object with the correct data for the chosen facade
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$build_deanalyse(contrasts, default_model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contrasts}}{named character vector of contrast definitions}

\item{\code{default_model}}{facade registry key, or NULL to read from config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DEAnalyse R6 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteinDataPrep-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProteinDataPrep-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinDataPrep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
