% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{R6_extract_values}
\alias{R6_extract_values}
\title{extract R6 class fields as list}
\usage{
R6_extract_values(r6class)
}
\description{
extract R6 class fields as list
}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ExternalReader}},
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{make_DEA_config_R6}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
