% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{read_annotation}
\alias{read_annotation}
\title{read annotation files}
\usage{
read_annotation(
  dsf,
  repeated = TRUE,
  SAINT = FALSE,
  prefix = "G_",
  QC = FALSE,
  shorten_sample_names = TRUE,
  sample_name_suffix_length = 14L,
  sample_name_display_column = "sampleName"
)
}
\arguments{
\item{dsf}{annotation table}

\item{repeated}{is this a repeated measurement}

\item{SAINT}{is this a SAINTexpress analysis}

\item{prefix}{prefix for group levels}

\item{QC}{if TRUE, read as QC annotation}

\item{shorten_sample_names}{derive short display sample names from long names}

\item{sample_name_suffix_length}{suffix length used for derived sample names}

\item{sample_name_display_column}{preferred derived display sample-name column}
}
\value{
list with annot (annotation table), atable (analtysis table configuration), contrasts list with contrasts.
}
\description{
read annotation files
}
\examples{
annot <- data.frame(
file = c("a1.raw","a2.raw","a3.raw","a4.raw"),
name = c("aa","ba","aa","ba"),
group = c("a","a","b","b"))
read_annotation(annot, QC = TRUE)

}
