% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{read_annotation_file}
\alias{read_annotation_file}
\title{read dataset file in csv, tsv or xlsx format}
\usage{
read_annotation_file(file_path)
}
\description{
read dataset file in csv, tsv or xlsx format
}
