% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Configuration.R
\name{read_BF_yamlR6}
\alias{read_BF_yamlR6}
\title{read yaml file}
\usage{
read_BF_yamlR6(yfile, application = "FragPipeTMT")
}
\value{
list with applications parameters
}
\description{
read yaml file
}
\examples{

yfile <- file.path(find.package("prolfquapp") , "/extdata/DIANN_BF/config1.yaml")
read_BF_yamlR6(yfile)

}
