% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_default.R,
%   R/proprocess_BGS_default.R
\name{read_BGS}
\alias{read_BGS}
\title{get BGS and fasta file location in folder}
\usage{
read_BGS(file = "Experiment1_Report_BGS Factory Report (Normal).tsv")

read_BGS(file = "Experiment1_Report_BGS Factory Report (Normal).tsv")
}
\value{
list with paths to data and fasta

list with paths to data and fasta
}
\description{
get BGS and fasta file location in folder

get BGS and fasta file location in folder
}
