% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_other_outputs.R
\name{read_combined_STY_file}
\alias{read_combined_STY_file}
\title{reads combined_site_STY file and converts to long format.}
\usage{
read_combined_STY_file(files)
}
\value{
tidy data table
}
\description{
reads combined_site_STY file and converts to long format.
}
