% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{read_table_data}
\alias{read_table_data}
\title{read dataset file in csv, tsv or xlsx format}
\usage{
read_table_data(file_path)
}
\description{
read dataset file in csv, tsv or xlsx format
}
