% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{read_yaml_deprec}
\alias{read_yaml_deprec}
\title{read yaml file}
\usage{
read_yaml_deprec(ymlfile, application = "FragPipeTMT")
}
\value{
list with applications parameters
}
\description{
read yaml file
}
