% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{sanitize_CD_subset_name}
\alias{sanitize_CD_subset_name}
\title{Sanitize a CompoundDiscoverer subset name for file paths}
\usage{
sanitize_CD_subset_name(x)
}
\arguments{
\item{x}{subset name}
}
\value{
filesystem-safe subset name
}
\description{
Sanitize a CompoundDiscoverer subset name for file paths
}
