% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{sanitize_grouping_var}
\alias{sanitize_grouping_var}
\title{Sanitize grouping variable in annotation file}
\usage{
sanitize_grouping_var(annot)
}
\description{
Sanitize grouping variable in annotation file
}
