% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_improtFrom.R
\name{set_lib_path}
\alias{set_lib_path}
\title{set library path with logging}
\usage{
set_lib_path(lib_path)
}
\arguments{
\item{lib_path}{path to R library directory}
}
\description{
set library path with logging
}
