% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinAnnotation.R
\name{sim_data_protAnnot}
\alias{sim_data_protAnnot}
\title{simulate peptdata and fitting protein annotation for testing}
\usage{
sim_data_protAnnot(Nprot = 100, PROTEIN = FALSE)
}
\arguments{
\item{Nprot}{number of proteins to simulate}

\item{PROTEIN}{if TRUE simulate protein-level data}
}
\description{
simulate peptdata and fitting protein annotation for testing
}
\examples{

res <- sim_data_protAnnot()
res <- sim_data_protAnnot(PROTEIN = TRUE)

}
