% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinAnnotationR6.R
\name{sim_peptdata_protAnnot}
\alias{sim_peptdata_protAnnot}
\title{simulate peptdata and fitting protein annotation for testing}
\usage{
sim_peptdata_protAnnot(Nprot = 100)
}
\description{
simulate peptdata and fitting protein annotation for testing
}
\examples{

res <- sim_peptdata_protAnnot()

}
