% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{strip_rownames}
\alias{strip_rownames}
\title{Strip pattern from row names of a matrix or data.frame}
\usage{
strip_rownames(.data, strip = "~lfq~light$")
}
\arguments{
\item{.data}{a matrix or data.frame with row names}

\item{strip}{regex pattern to remove from row names}
}
\value{
the input with cleaned row names
}
\description{
Strip pattern from row names of a matrix or data.frame
}
