% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sync_opt_config}
\alias{sync_opt_config}
\title{Synchronize opt and config}
\usage{
sync_opt_config(opt, config)
}
\description{
Synchronize opt and config
}
\examples{
opt <- list()
config <- make_DEA_config_R6()
xx <- sync_opt_config(opt, config)
stopifnot(names(xx$opt) \%in\% c("software","outdir"))
stopifnot(xx$opt$software == "DIANN")
stopifnot(xx$opt$outdir == ".")
opt$software <- "FP_TMT"
opt$outdir <- "testdir"
xx <- sync_opt_config(opt, config)
stopifnot(xx$config$path == "testdir")
stopifnot(xx$config$software == "FP_TMT")

}
