% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{tidy_FragPipe_psm_V2}
\alias{tidy_FragPipe_psm_V2}
\title{read psm.tsv produced by FragPipe and convert into long format}
\usage{
tidy_FragPipe_psm_V2(
  psm_files,
  purity_threshold = 0.5,
  PeptideProphetProb = 0.9,
  abundance_threshold = 0,
  quan_column_prefix = "^Intensity",
  aggregate = TRUE
)
}
\arguments{
\item{psm_files}{path(s) to psm.tsv file(s)}

\item{purity_threshold}{purity threshold default = 0.5}

\item{PeptideProphetProb}{default 0.9}

\item{abundance_threshold}{minimum abundance threshold}

\item{quan_column_prefix}{regex prefix for quantitative columns}

\item{aggregate}{aggregate spectra to psm level}
}
\description{
read psm.tsv produced by FragPipe and convert into long format
}
