% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{tidy_mzMineFeatures}
\alias{tidy_mzMineFeatures}
\title{convert mzmine features to tidy table}
\usage{
tidy_mzMineFeatures(data)
}
\arguments{
\item{data}{path to csv or data frame of mzMine features}
}
\value{
A long-format tibble with one row per feature and data file.
}
\description{
convert mzmine features to tidy table
}
\examples{
if(FALSE){
raw_df <- readr::read_csv("outputs-20250407T1707/mzmine/result_features.csv")
res <- tidy_mzMineFeatures(raw_df)
head(res)
}
}
