% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{tidyMQ_Peptides}
\alias{tidyMQ_Peptides}
\title{parse MQ peptides.txt}
\usage{
tidyMQ_Peptides(MQPeptides, proteotypic_only = TRUE)
}
\arguments{
\item{MQPeptides}{data.frame generated with read.csv("peptide.txt",sep = "\\t", stringsAsFactors = FALSE)}
}
\description{
parse MQ peptides.txt
}
\examples{
peptide_txt <- prolfqua::find_package_file("prolfquapp", "samples/maxquant_txt/tiny2.zip")
peptides_txt <- read.csv(
  unz(peptide_txt, "peptides.txt"),
  header = TRUE, stringsAsFactors = FALSE, sep = "\t")
mq_peptides <- tidyMQ_Peptides(peptides_txt)
}
\seealso{
Other MaxQuant: 
\code{\link{MaxQuant}},
\code{\link{dataset_template_MAXQUANT}()},
\code{\link{tidyMQ_Evidence}()},
\code{\link{tidyMQ_ProteinGroups}()}
}
\concept{MaxQuant}
\keyword{internal}
