% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{tidyMQ_ProteinGroups}
\alias{tidyMQ_ProteinGroups}
\title{extract intensities and annotations from MQ proteinGroups.txt}
\usage{
tidyMQ_ProteinGroups(MQProteinGroups)
}
\arguments{
\item{MQProteinGroups}{data.frame generated with read.csv("peptide.txt",sep="\\t", stringsAsFactors=FALSE)}
}
\description{
extract intensities and annotations from MQ proteinGroups.txt
}
\examples{
protein_txt <- prolfqua::find_package_file("prolfquapp","samples/maxquant_txt/tiny2.zip")
protein_txt <- read.csv(
  unz(protein_txt, "proteinGroups.txt"),
  header = TRUE, stringsAsFactors = FALSE, sep = "\t")
mq_proteins <-tidyMQ_ProteinGroups(protein_txt)
}
\seealso{
Other MaxQuant: 
\code{\link{MaxQuant}},
\code{\link{dataset_template_MAXQUANT}()},
\code{\link{tidyMQ_Evidence}()},
\code{\link{tidyMQ_Peptides}()}
}
\concept{MaxQuant}
\keyword{internal}
