% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_lfqdata.R
\name{transform_lfqdata}
\alias{transform_lfqdata}
\title{transform lfq data using robscale, vsn or log2, Assumes that data is not transformed (still needs log2 transformation)}
\usage{
transform_lfqdata(lfqdata, method = c("robscale", "vsn", "none", "log2"))
}
\arguments{
\item{lfqdata}{\code{\link[prolfqua]{LFQData}}}

\item{method}{normalization method to use}
}
\description{
transform lfq data using robscale, vsn or log2, Assumes that data is not transformed (still needs log2 transformation)
}
\examples{
istar <- prolfqua::prolfqua_data('data_ionstar')$filtered()
config <- prolfqua:::old2new(istar$config)
tmp <- prolfqua::LFQData$new(istar$data, config)
tmp2 <- transform_lfqdata(tmp)

}
