% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_lfqdata.R
\name{validate_prolfquapp_anndata}
\alias{validate_prolfquapp_anndata}
\title{Validate that an AnnData has the prolfquapp uns namespace}
\usage{
validate_prolfquapp_anndata(adata)
}
\arguments{
\item{adata}{an AnnData object}
}
\description{
Validate that an AnnData has the prolfquapp uns namespace
}
