% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{write_DEA}
\alias{write_DEA}
\title{Write differential expression analysis results}
\usage{
write_DEA(
  GRP2,
  outpath,
  ORA = TRUE,
  GSEA = TRUE,
  xlsxname = "AnalysisResults",
  write = TRUE
)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}

\item{outpath}{path to place output}

\item{ORA}{if TRUE write ORA gene lists}

\item{GSEA}{if TRUE write GSEA rank files}

\item{xlsxname}{file name for xlsx}

\item{write}{if TRUE write files to disk}
}
\description{
Write differential expression analysis results
}
\seealso{
Other workflow: 
\code{\link{make_DEA_report}()},
\code{\link{make_SummarizedExperiment}()}
}
\concept{workflow}
