% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{write_index_html}
\alias{write_index_html}
\title{write index.html file with links to all relevant files:}
\usage{
write_index_html(file_path_list, result_dir)
}
\arguments{
\item{file_path_list}{named list of output file paths}

\item{result_dir}{directory for the index.html output}
}
\description{
write index.html file with links to all relevant files:
}
\examples{
.resdir <- "."
write_index_html(prolfquapp:::.test_links,tempdir())
}
