% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{add_contrasts_vec}
\alias{add_contrasts_vec}
\title{add vector of contrasts to annotation data frame}
\usage{
add_contrasts_vec(xx, Contrasts)
}
\arguments{
\item{xx}{annotation data frame}

\item{Contrasts}{character vector of contrasts}
}
\description{
add vector of contrasts to annotation data frame
}
\examples{
annot <- data.frame(Group = rep(c("A","B","C"), each = 3))
annot$Name
}
