% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinAnnotation.R
\name{add_RevCon}
\alias{add_RevCon}
\title{add REV and zz entries - used for testing}
\usage{
add_RevCon(stringsAll, pattern_decoys = "REV_", pattern_contaminants = "zz")
}
\description{
add REV and zz entries - used for testing
}
