% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{build_protein_annot}
\alias{build_protein_annot}
\title{Dataset protein annot}
\usage{
build_protein_annot(
  lfqdata,
  msdata,
  idcol = c(protein_Id = "Protein.Group"),
  cleaned_protein_id = "Protein.Group.2",
  protein_description = "fasta.header",
  nr_children = "nrPeptides",
  more_columns = c("fasta.id")
)
}
\arguments{
\item{msdata}{data frame}

\item{more_columns}{more columns to include}

\item{idcolName}{name of column with ID's}

\item{protein_annot}{fasta haeder column}
}
\description{
Dataset protein annot
}
