% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinAnnotation.R
\name{build_protein_annot}
\alias{build_protein_annot}
\title{build Dataset protein annot, defaults are compatible with DIANN}
\usage{
build_protein_annot(
  lfqdata,
  msdata,
  idcol = c(protein_Id = "Protein.Group"),
  cleaned_protein_id = "Protein.Group.2",
  protein_description = "fasta.header",
  exp_nr_children = "nrPeptides",
  full_id = "fasta.id",
  more_columns = c("fasta.id"),
  pattern_contaminants = "^zz|^CON",
  pattern_decoys = "REV_"
)
}
\arguments{
\item{lfqdata}{LFQData}

\item{msdata}{data frame}

\item{more_columns}{more columns to include}

\item{idcolName}{name of column with ID's}

\item{protein_annot}{fasta haeder column}
}
\description{
build Dataset protein annot, defaults are compatible with DIANN
}
\examples{
# example code

}
