% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_IBAQ.R
\name{compute_IBAQ_values}
\alias{compute_IBAQ_values}
\title{compute IBAQ values}
\usage{
compute_IBAQ_values(
  lfqdata,
  protein_annotation,
  protein_length = "protein_length",
  nr_tryptic_peptides = "nr_tryptic_peptides"
)
}
\description{
compute IBAQ values
}
\examples{
pAlf <- sim_data_protAnnot()
xd <- compute_IBAQ_values(pAlf$lfqdata, pAlf$pannot)
xd$response()
pAlf <- sim_data_protAnnot(PROTEIN = TRUE)
xd <- compute_IBAQ_values(pAlf$lfqdata, pAlf$pannot)
xd$response()

}
