% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_protein_annot}
\alias{dataset_protein_annot}
\title{dataset protein annot}
\usage{
dataset_protein_annot(
  msdata,
  idcolName,
  proteinID = "protein_Id",
  protein_annot = "fasta.header"
)
}
\description{
dataset protein annot
}
