% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_protein_annot}
\alias{dataset_protein_annot}
\title{Dataset protein annot}
\usage{
dataset_protein_annot(
  msdata,
  idcol = c(protein_Id = "Protein.Group"),
  protein_annot = "fasta.header",
  more_columns = c("nrPeptides", "fasta.id")
)
}
\arguments{
\item{msdata}{data frame}

\item{protein_annot}{fasta haeder column}

\item{more_columns}{more columns to include}

\item{idcolName}{name of column with ID's}
}
\description{
Dataset protein annot
}
