% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_protein_annot}
\alias{dataset_protein_annot}
\title{dataset protein annot}
\usage{
dataset_protein_annot(peptide, atable, protein_annot = "Protein.Description")
}
\description{
dataset protein annot
}
