% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{dataset_protein_annot}
\alias{dataset_protein_annot}
\title{Dataset protein annot}
\usage{
dataset_protein_annot(
  msdata,
  idcol = c(protein_Id = "Protein.Group"),
  protein_annot = "fasta.header",
  more_columns = c("nrPeptides", "fasta.id")
)
}
\arguments{
\item{msdata}{data frame}

\item{idcol}{named vector mapping protein ID column}

\item{protein_annot}{fasta header column name}

\item{more_columns}{more columns to include}
}
\description{
Dataset protein annot
}
