% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{dataset_template_FP_TMT}
\alias{dataset_template_FP_TMT}
\title{get dataset annotation template}
\usage{
dataset_template_FP_TMT(files)
}
\arguments{
\item{files}{list with data and fasta file paths}
}
\value{
data.frame
}
\description{
get dataset annotation template
}
