% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_MaxQuant.R
\name{dataset_template_MAXQUANT}
\alias{dataset_template_MAXQUANT}
\title{create template dataset for MAXQUANT data}
\usage{
dataset_template_MAXQUANT(files)
}
\description{
create template dataset for MAXQUANT data
}
