% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEAnalyse.R
\name{DEAnalyse}
\alias{DEAnalyse}
\title{Differential expression analysis engine}
\description{
Differential expression analysis engine

Differential expression analysis engine
}
\details{
Takes a prepared \code{ProteinDataPrep} object and runs statistical modelling:
fits linear and GLM models, computes moderated contrasts, merges results.
}
\examples{
pep <- prolfqua::sim_lfq_data_peptide_config(Nprot = 100)
pep <- prolfqua::LFQData$new(pep$data, pep$config)
pA <- data.frame(protein_Id = unique(pep$data$protein_Id))
pA <- pA |> dplyr::mutate(fasta.annot = paste0(pA$protein_Id, "_description"))
pA <- prolfquapp::ProteinAnnotation$new(pep, row_annot = pA, description = "fasta.annot")
GRP2 <- prolfquapp::make_DEA_config_R6()
GRP2$processing_options$diff_threshold <- 0.2
GRP2$processing_options$transform <- "robscale"

contrasts <- c("AVsC" = "group_A - group_Ctrl", BVsC = "group_B - group_Ctrl")

data_prep <- prolfquapp::ProteinDataPrep$new(pep, pA, GRP2)
data_prep$cont_decoy_summary()
data_prep$remove_cont_decoy()
data_prep$aggregate()
data_prep$transform_data()

deanalyse <- prolfquapp::DEAnalyse$new(data_prep, contrasts)
mod <- deanalyse$build_model_linear_protein()
contlm <- deanalyse$get_contrasts_linear_protein()
merged <- deanalyse$get_contrasts_merged_protein()
stopifnot(nrow(merged$get_contrasts()) == 200)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{lfq_data_peptide}}{LFQData peptide level}

\item{\code{lfq_data}}{LFQData protein level}

\item{\code{lfq_data_transformed}}{normalized LFQData}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{contrasts}}{vector with contrasts}

\item{\code{FDR_threshold}}{FDR threshold}

\item{\code{diff_threshold}}{difference threshold}

\item{\code{summary}}{data.frame with contaminant/decoy summary}

\item{\code{annotated_contrasts}}{contrasts joined with row annotations}

\item{\code{annotated_contrasts_signif}}{significant annotated contrasts}

\item{\code{formula}}{model formula}

\item{\code{models}}{list of fitted models}

\item{\code{contrast_results}}{list of contrast results}

\item{\code{m1_linear}}{Linear_Model}

\item{\code{m2_missing}}{Imputed_Mean}

\item{\code{m3_merged}}{mergedModel}

\item{\code{m4_glm_protein}}{glmModel}

\item{\code{m4_glm_peptide}}{glmModelPeptide}

\item{\code{default_model}}{default model key}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAnalyse-new}{\code{DEAnalyse$new()}}
\item \href{#method-DEAnalyse-create_model_formula}{\code{DEAnalyse$create_model_formula()}}
\item \href{#method-DEAnalyse-build_model_linear_protein}{\code{DEAnalyse$build_model_linear_protein()}}
\item \href{#method-DEAnalyse-get_strategy_glm_prot}{\code{DEAnalyse$get_strategy_glm_prot()}}
\item \href{#method-DEAnalyse-build_model_glm_protein}{\code{DEAnalyse$build_model_glm_protein()}}
\item \href{#method-DEAnalyse-build_model_glm_peptide}{\code{DEAnalyse$build_model_glm_peptide()}}
\item \href{#method-DEAnalyse-get_contrasts_linear_protein}{\code{DEAnalyse$get_contrasts_linear_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_glm_peptide}{\code{DEAnalyse$get_contrasts_glm_peptide()}}
\item \href{#method-DEAnalyse-get_contrasts_glm_protein}{\code{DEAnalyse$get_contrasts_glm_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_missing_protein}{\code{DEAnalyse$get_contrasts_missing_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_merged_protein}{\code{DEAnalyse$get_contrasts_merged_protein()}}
\item \href{#method-DEAnalyse-get_annotated_contrasts}{\code{DEAnalyse$get_annotated_contrasts()}}
\item \href{#method-DEAnalyse-filter_contrasts}{\code{DEAnalyse$filter_contrasts()}}
\item \href{#method-DEAnalyse-clone}{\code{DEAnalyse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-new}{}}}
\subsection{Method \code{new()}}{
Initialize DEAnalyse from a prepared ProteinDataPrep object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$new(data_prep, contrasts, default_model = "mergedModel")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_prep}}{ProteinDataPrep object with aggregated and normalized data}

\item{\code{contrasts}}{named vector of contrast definitions}

\item{\code{default_model}}{which model to use for final results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-create_model_formula"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-create_model_formula}{}}}
\subsection{Method \code{create_model_formula()}}{
Create model formula from transformed data config
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$create_model_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_linear_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_linear_protein}{}}}
\subsection{Method \code{build_model_linear_protein()}}{
Fit linear model at protein level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_linear_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_strategy_glm_prot"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_strategy_glm_prot}{}}}
\subsection{Method \code{get_strategy_glm_prot()}}{
Get GLM strategy for protein-level missingness model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_strategy_glm_prot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_protein}{}}}
\subsection{Method \code{build_model_glm_protein()}}{
Fit generalized linear model at protein level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_peptide}{}}}
\subsection{Method \code{build_model_glm_peptide()}}{
Fit generalized linear model at peptide level (not yet implemented)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_linear_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_linear_protein}{}}}
\subsection{Method \code{get_contrasts_linear_protein()}}{
Compute moderated contrasts from linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_linear_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_glm_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_glm_peptide}{}}}
\subsection{Method \code{get_contrasts_glm_peptide()}}{
Compute moderated contrasts from GLM peptide model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_glm_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_glm_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_glm_protein}{}}}
\subsection{Method \code{get_contrasts_glm_protein()}}{
Compute moderated contrasts from GLM protein model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_glm_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_missing_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_missing_protein}{}}}
\subsection{Method \code{get_contrasts_missing_protein()}}{
Compute moderated contrasts from missing-value imputation model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_missing_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_merged_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_merged_protein}{}}}
\subsection{Method \code{get_contrasts_merged_protein()}}{
Merge linear and missing-value contrasts (or use linear only if model_missing = FALSE)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_merged_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_annotated_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_annotated_contrasts}{}}}
\subsection{Method \code{get_annotated_contrasts()}}{
Join default model contrasts with protein row annotations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_annotated_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_contrasts}{}}}
\subsection{Method \code{filter_contrasts()}}{
Return contrast rows passing FDR and difference thresholds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
