% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_report.R
\name{DEAnalyse}
\alias{DEAnalyse}
\title{will replace make_DEA_report}
\description{
will replace make_DEA_report

will replace make_DEA_report
}
\examples{
# example code

pep <- prolfqua::sim_lfq_data_protein_config(Nprot = 100)
pep <- prolfqua::LFQData$new(pep$data, pep$config)
pA <- data.frame(protein_Id = unique(pep$data$protein_Id))
pA <- pA |> dplyr::mutate(fasta.annot = paste0(pA$protein_Id, "_description"))
pA <- prolfqua::ProteinAnnotation$new(pep,row_annot = pA ,description = "fasta.annot")
GRP2 <- prolfquapp::make_DEA_config_R6()
GRP2$processing_options$transform <- "robscale"
pep$factors()
GRP2$pop <- list(Contrasts = c("AVsC" = "group_A - group_Ctrl", BVsC = "group_B - group_Ctrl"))
deanalyse <- DEAnalyse$new(pep, pA, GRP2)
deanalyse$cont_decoy_summary()
deanalyse$GRP2$processing_options$remove_cont = TRUE
deanalyse$remove_cont_decoy()
deanalyse$transform_data()
deanalyse$create_model_formula()
deanalyse$build_model()
deanalyse$compute_contrasts()
deanalyse$filter_contrasts()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lfqdata}}{`LFQData`}

\item{\code{rowAnnot}}{`ProteinAnnotation`}

\item{\code{GRP2}}{`ProlfquAppConfig`
initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAnalyse-new}{\code{DEAnalyse$new()}}
\item \href{#method-DEAnalyse-cont_decoy_summary}{\code{DEAnalyse$cont_decoy_summary()}}
\item \href{#method-DEAnalyse-remove_cont_decoy}{\code{DEAnalyse$remove_cont_decoy()}}
\item \href{#method-DEAnalyse-transform_data}{\code{DEAnalyse$transform_data()}}
\item \href{#method-DEAnalyse-create_model_formula}{\code{DEAnalyse$create_model_formula()}}
\item \href{#method-DEAnalyse-build_model}{\code{DEAnalyse$build_model()}}
\item \href{#method-DEAnalyse-compute_contrasts}{\code{DEAnalyse$compute_contrasts()}}
\item \href{#method-DEAnalyse-filter_contrasts}{\code{DEAnalyse$filter_contrasts()}}
\item \href{#method-DEAnalyse-clone}{\code{DEAnalyse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$new(lfqdata, rowAnnot, GRP2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lfqdata}}{lfqdata}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{GRP2}}{ProlfquAppConfig}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-cont_decoy_summary"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-cont_decoy_summary}{}}}
\subsection{Method \code{cont_decoy_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$cont_decoy_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-remove_cont_decoy"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-remove_cont_decoy}{}}}
\subsection{Method \code{remove_cont_decoy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$remove_cont_decoy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-create_model_formula"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-create_model_formula}{}}}
\subsection{Method \code{create_model_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$create_model_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model}{}}}
\subsection{Method \code{build_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-compute_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-compute_contrasts}{}}}
\subsection{Method \code{compute_contrasts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$compute_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_contrasts}{}}}
\subsection{Method \code{filter_contrasts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
