% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalyseR6.R
\name{DEAnalyse}
\alias{DEAnalyse}
\title{will replace make_DEA_report}
\description{
will replace make_DEA_report

will replace make_DEA_report
}
\examples{
# example code

pep <- prolfqua::sim_lfq_data_protein_config(Nprot = 100)

pep <- prolfqua::LFQData$new(pep$data, pep$config)
pA <- data.frame(protein_Id = unique(pep$data$protein_Id))
pA <- pA |> dplyr::mutate(fasta.annot = paste0(pA$protein_Id, "_description"))
pA <- prolfquapp::ProteinAnnotation$new(pep,row_annot = pA ,description = "fasta.annot")
GRP2 <- prolfquapp::make_DEA_config_R6()
GRP2$processing_options$diff_threshold = 0.2

GRP2$processing_options$transform <- "robscale"
pep$factors()
contrasts = c("AVsC" = "group_A - group_Ctrl", BVsC = "group_B - group_Ctrl")
# DEAnalyse$debug("filter_contrasts")
deanalyse <- DEAnalyse$new(pep, pA, GRP2, contrasts)
deanalyse$cont_decoy_summary()
deanalyse$prolfq_app_config$processing_options$remove_cont = TRUE
deanalyse$remove_cont_decoy()
deanalyse$transform_data()
deanalyse$create_model_formula()
mod <- deanalyse$build_model_linear()
contlm <- deanalyse$get_contrasts_linear()

deanalyse$filter_contrasts()

xd <- deanalyse$filter_data()
xd <- deanalyse$contrasts_to_Grob()
bb <- deanalyse$get_boxplots()
bx <- deanalyse$get_boxplots_contrasts()
grid::grid.draw(bx$bxpl_grobs[[1]])
# deanalyse$write_boxplots_contrasts("test.pdf")
deanalyse$build_model_glm_protein()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lfqdata}}{LFQData}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{contrasts}}{vector with contrasts}

\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{reference_proteins}}{reference proteins to use for internal normalization}

\item{\code{decoy_summary}}{decoy_summary}

\item{\code{lfqData}}{lfqData}

\item{\code{lfqData_transformed}}{todo}

\item{\code{lfqData_subset}}{todo}

\item{\code{formula}}{todo}

\item{\code{models}}{todo}

\item{\code{contrastRes}}{todo}

\item{\code{FDR_threshold}}{fdr threshold}

\item{\code{diff_threshold}}{diff_threshold}

\item{\code{m1_linear}}{linearModel}

\item{\code{m2_missing}}{imputedModel}

\item{\code{m3_merged}}{mergedModel}

\item{\code{m4_glm_protein}}{m4_glm_protein}

\item{\code{m4_glm_peptide}}{m4_glm_peptide}

\item{\code{default_model}}{default_model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAnalyse-new}{\code{DEAnalyse$new()}}
\item \href{#method-DEAnalyse-cont_decoy_summary}{\code{DEAnalyse$cont_decoy_summary()}}
\item \href{#method-DEAnalyse-remove_cont_decoy}{\code{DEAnalyse$remove_cont_decoy()}}
\item \href{#method-DEAnalyse-transform_data}{\code{DEAnalyse$transform_data()}}
\item \href{#method-DEAnalyse-create_model_formula}{\code{DEAnalyse$create_model_formula()}}
\item \href{#method-DEAnalyse-create_model_formula_peptide}{\code{DEAnalyse$create_model_formula_peptide()}}
\item \href{#method-DEAnalyse-build_model_linear}{\code{DEAnalyse$build_model_linear()}}
\item \href{#method-DEAnalyse-build_model_glm_protein}{\code{DEAnalyse$build_model_glm_protein()}}
\item \href{#method-DEAnalyse-build_model_glm_peptide}{\code{DEAnalyse$build_model_glm_peptide()}}
\item \href{#method-DEAnalyse-get_contrasts_linear}{\code{DEAnalyse$get_contrasts_linear()}}
\item \href{#method-DEAnalyse-get_contrasts_glm}{\code{DEAnalyse$get_contrasts_glm()}}
\item \href{#method-DEAnalyse-get_contrasts_missing}{\code{DEAnalyse$get_contrasts_missing()}}
\item \href{#method-DEAnalyse-get_contrasts_merged}{\code{DEAnalyse$get_contrasts_merged()}}
\item \href{#method-DEAnalyse-filter_contrasts}{\code{DEAnalyse$filter_contrasts()}}
\item \href{#method-DEAnalyse-filter_data}{\code{DEAnalyse$filter_data()}}
\item \href{#method-DEAnalyse-get_boxplots}{\code{DEAnalyse$get_boxplots()}}
\item \href{#method-DEAnalyse-contrasts_to_Grob}{\code{DEAnalyse$contrasts_to_Grob()}}
\item \href{#method-DEAnalyse-get_boxplots_contrasts}{\code{DEAnalyse$get_boxplots_contrasts()}}
\item \href{#method-DEAnalyse-write_boxplots_contrasts}{\code{DEAnalyse$write_boxplots_contrasts()}}
\item \href{#method-DEAnalyse-clone}{\code{DEAnalyse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$new(
  lfqdata,
  rowAnnot,
  prolfq_app_config,
  contrasts,
  default_model = "mergedModel"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lfqdata}}{lfqdata}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{contrasts}}{vector with contrasts}

\item{\code{FDR_threshold}}{FDR_threshold}

\item{\code{diff_threshold}}{diff_threshold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-cont_decoy_summary"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-cont_decoy_summary}{}}}
\subsection{Method \code{cont_decoy_summary()}}{
count number of decoys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$cont_decoy_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-remove_cont_decoy"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-remove_cont_decoy}{}}}
\subsection{Method \code{remove_cont_decoy()}}{
remove contaminants and decoys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$remove_cont_decoy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
transform data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-create_model_formula"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-create_model_formula}{}}}
\subsection{Method \code{create_model_formula()}}{
static create model formula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$create_model_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-create_model_formula_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-create_model_formula_peptide}{}}}
\subsection{Method \code{create_model_formula_peptide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$create_model_formula_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_linear"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_linear}{}}}
\subsection{Method \code{build_model_linear()}}{
fit linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_linear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_protein}{}}}
\subsection{Method \code{build_model_glm_protein()}}{
fit generalized linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_peptide}{}}}
\subsection{Method \code{build_model_glm_peptide()}}{
fit generalized linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_linear"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_linear}{}}}
\subsection{Method \code{get_contrasts_linear()}}{
compute contrasts linear
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_linear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_glm"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_glm}{}}}
\subsection{Method \code{get_contrasts_glm()}}{
get contrasts from glm model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_glm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_missing"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_missing}{}}}
\subsection{Method \code{get_contrasts_missing()}}{
compute missing contrasts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_missing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_merged"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_merged}{}}}
\subsection{Method \code{get_contrasts_merged()}}{
merge contrasts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_merged()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_contrasts}{}}}
\subsection{Method \code{filter_contrasts()}}{
filter contrasts for threshold
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_data"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_data}{}}}
\subsection{Method \code{filter_data()}}{
filter transformed lfq data for significant proteins.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_boxplots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_boxplots}{}}}
\subsection{Method \code{get_boxplots()}}{
create boxplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_boxplots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-contrasts_to_Grob"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-contrasts_to_Grob}{}}}
\subsection{Method \code{contrasts_to_Grob()}}{
create boxplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$contrasts_to_Grob()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_boxplots_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_boxplots_contrasts}{}}}
\subsection{Method \code{get_boxplots_contrasts()}}{
get box with contrast information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_boxplots_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-write_boxplots_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-write_boxplots_contrasts}{}}}
\subsection{Method \code{write_boxplots_contrasts()}}{
write boxplots contrasts to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$write_boxplots_contrasts(filename = "boxplots")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename to write to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
