% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEAnalyse.R
\name{DEAnalyse}
\alias{DEAnalyse}
\title{will replace make_DEA_report}
\description{
will replace make_DEA_report

will replace make_DEA_report
}
\examples{
# example code

pep <- prolfqua::sim_lfq_data_peptide_config(Nprot = 100)

pep <- prolfqua::LFQData$new(pep$data, pep$config)
pA <- data.frame(protein_Id = unique(pep$data$protein_Id))
pA <- pA |> dplyr::mutate(fasta.annot = paste0(pA$protein_Id, "_description"))
pA <- prolfquapp::ProteinAnnotation$new(pep,row_annot = pA ,description = "fasta.annot")
GRP2 <- prolfquapp::make_DEA_config_R6()
GRP2$processing_options$diff_threshold = 0.2

GRP2$processing_options$transform <- "robscale"
pep$factors()
contrasts = c("AVsC" = "group_A - group_Ctrl", BVsC = "group_B - group_Ctrl")
#DEAnalyse$debug("get_contrasts_glm_peptide")
#DEAnalyse$debug("build_model_glm_protein")
deanalyse <- prolfquapp::DEAnalyse$new(pep, pA, GRP2, contrasts)
deanalyse$lfq_data_peptide$hierarchy_counts()
deanalyse$cont_decoy_summary()
deanalyse$prolfq_app_config$processing_options$remove_cont = TRUE
deanalyse$remove_cont_decoy()
deanalyse$aggregate()
pl <- deanalyse$get_aggregation_plots(exp_nr_children = 10)
print(pl$plots[[3]])
deanalyse$transform_data()
mod <- deanalyse$build_model_linear_protein()
contlm <- deanalyse$get_contrasts_linear_protein()


merged <- deanalyse$get_contrasts_merged_protein()
stopifnot(nrow(merged$get_contrasts()) == 200)
#deanalyse$create_model_formula()
#deanalyse$build_model_glm_protein()
#deanalyse$build_model_glm_peptide()
xprot <- deanalyse$get_contrasts_glm_protein()
stopifnot(nrow(merged$get_contrasts()) == 200)
xpep <- deanalyse$get_contrasts_glm_peptide()
xprot$get_contrasts()
xprot$get_Plotter()$volcano()
xpep$get_Plotter()$volcano()
sr <- deanalyse$lfq_data_peptide$get_Summariser()



deanalyse$filter_contrasts()

xd <- deanalyse$filter_data()
xd <- deanalyse$contrasts_to_Grob()
bb <- deanalyse$get_boxplots()
bx <- deanalyse$get_boxplots_contrasts()
grid::grid.draw(bx$bxpl_grobs[[1]])
# deanalyse$write_boxplots_contrasts("test.pdf")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{lfq_data_peptide}}{LFQData peptide level}

\item{\code{lfq_data}}{LFQData}

\item{\code{lfq_data_transformed}}{todo}

\item{\code{lfq_data_subset}}{todo}

\item{\code{aggregator}}{aggregator}

\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{contrasts}}{vector with contrasts}

\item{\code{FDR_threshold}}{fdr threshold}

\item{\code{diff_threshold}}{diff_threshold}

\item{\code{reference_proteins}}{reference proteins to use for internal normalization}

\item{\code{formula}}{todo}

\item{\code{formula}}{glm peptide todo}

\item{\code{models}}{todo}

\item{\code{contrast_results}}{todo}

\item{\code{m1_linear}}{linearModel}

\item{\code{m2_missing}}{imputedModel}

\item{\code{m3_merged}}{mergedModel}

\item{\code{m4_glm_protein}}{m4_glm_protein}

\item{\code{m4_glm_peptide}}{m4_glm_peptide}

\item{\code{default_model}}{default_model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAnalyse-new}{\code{DEAnalyse$new()}}
\item \href{#method-DEAnalyse-cont_decoy_summary}{\code{DEAnalyse$cont_decoy_summary()}}
\item \href{#method-DEAnalyse-remove_cont_decoy}{\code{DEAnalyse$remove_cont_decoy()}}
\item \href{#method-DEAnalyse-aggregate}{\code{DEAnalyse$aggregate()}}
\item \href{#method-DEAnalyse-get_aggregation_plots}{\code{DEAnalyse$get_aggregation_plots()}}
\item \href{#method-DEAnalyse-write_aggregation_plots}{\code{DEAnalyse$write_aggregation_plots()}}
\item \href{#method-DEAnalyse-transform_data}{\code{DEAnalyse$transform_data()}}
\item \href{#method-DEAnalyse-create_model_formula}{\code{DEAnalyse$create_model_formula()}}
\item \href{#method-DEAnalyse-build_model_linear_protein}{\code{DEAnalyse$build_model_linear_protein()}}
\item \href{#method-DEAnalyse-build_model_glm_protein}{\code{DEAnalyse$build_model_glm_protein()}}
\item \href{#method-DEAnalyse-build_model_glm_peptide}{\code{DEAnalyse$build_model_glm_peptide()}}
\item \href{#method-DEAnalyse-get_contrasts_linear_protein}{\code{DEAnalyse$get_contrasts_linear_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_glm_peptide}{\code{DEAnalyse$get_contrasts_glm_peptide()}}
\item \href{#method-DEAnalyse-get_contrasts_glm_protein}{\code{DEAnalyse$get_contrasts_glm_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_missing_protein}{\code{DEAnalyse$get_contrasts_missing_protein()}}
\item \href{#method-DEAnalyse-get_contrasts_merged_protein}{\code{DEAnalyse$get_contrasts_merged_protein()}}
\item \href{#method-DEAnalyse-filter_contrasts}{\code{DEAnalyse$filter_contrasts()}}
\item \href{#method-DEAnalyse-filter_data}{\code{DEAnalyse$filter_data()}}
\item \href{#method-DEAnalyse-get_boxplots}{\code{DEAnalyse$get_boxplots()}}
\item \href{#method-DEAnalyse-contrasts_to_Grob}{\code{DEAnalyse$contrasts_to_Grob()}}
\item \href{#method-DEAnalyse-get_boxplots_contrasts}{\code{DEAnalyse$get_boxplots_contrasts()}}
\item \href{#method-DEAnalyse-write_boxplots_contrasts}{\code{DEAnalyse$write_boxplots_contrasts()}}
\item \href{#method-DEAnalyse-clone}{\code{DEAnalyse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$new(
  lfq_data_peptide,
  rowAnnot,
  prolfq_app_config,
  contrasts,
  default_model = "mergedModel"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rowAnnot}}{ProteinAnnotation}

\item{\code{prolfq_app_config}}{ProlfquAppConfig}

\item{\code{contrasts}}{vector with contrasts}

\item{\code{lfq_data}}{lfq_data}

\item{\code{FDR_threshold}}{FDR_threshold}

\item{\code{diff_threshold}}{diff_threshold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-cont_decoy_summary"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-cont_decoy_summary}{}}}
\subsection{Method \code{cont_decoy_summary()}}{
count number of decoys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$cont_decoy_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-remove_cont_decoy"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-remove_cont_decoy}{}}}
\subsection{Method \code{remove_cont_decoy()}}{
remove contaminants and decoys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$remove_cont_decoy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
aggregate peptide data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$aggregate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_aggregation_plots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_aggregation_plots}{}}}
\subsection{Method \code{get_aggregation_plots()}}{
get aggregation plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_aggregation_plots(exp_nr_children = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp_nr_children}}{nr children to filter; default >=2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-write_aggregation_plots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-write_aggregation_plots}{}}}
\subsection{Method \code{write_aggregation_plots()}}{
write aggregation plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$write_aggregation_plots(exp_nr_children = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp_nr_children}}{nr children to filter; default >=2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
transform data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-create_model_formula"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-create_model_formula}{}}}
\subsection{Method \code{create_model_formula()}}{
create model formula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$create_model_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_linear_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_linear_protein}{}}}
\subsection{Method \code{build_model_linear_protein()}}{
fit linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_linear_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_protein}{}}}
\subsection{Method \code{build_model_glm_protein()}}{
fit generalized linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-build_model_glm_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-build_model_glm_peptide}{}}}
\subsection{Method \code{build_model_glm_peptide()}}{
fit generalized linear model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$build_model_glm_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_linear_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_linear_protein}{}}}
\subsection{Method \code{get_contrasts_linear_protein()}}{
compute contrasts linear
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_linear_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_glm_peptide"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_glm_peptide}{}}}
\subsection{Method \code{get_contrasts_glm_peptide()}}{
get contrasts from glm model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_glm_peptide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_glm_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_glm_protein}{}}}
\subsection{Method \code{get_contrasts_glm_protein()}}{
get contrasts from glm model for peptides
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_glm_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_missing_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_missing_protein}{}}}
\subsection{Method \code{get_contrasts_missing_protein()}}{
compute missing contrasts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_missing_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_contrasts_merged_protein"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_contrasts_merged_protein}{}}}
\subsection{Method \code{get_contrasts_merged_protein()}}{
merge contrasts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_contrasts_merged_protein()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_contrasts}{}}}
\subsection{Method \code{filter_contrasts()}}{
filter contrasts for threshold
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-filter_data"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-filter_data}{}}}
\subsection{Method \code{filter_data()}}{
filter transformed lfq data for significant proteins.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$filter_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_boxplots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_boxplots}{}}}
\subsection{Method \code{get_boxplots()}}{
create boxplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_boxplots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-contrasts_to_Grob"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-contrasts_to_Grob}{}}}
\subsection{Method \code{contrasts_to_Grob()}}{
create boxplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$contrasts_to_Grob()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-get_boxplots_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-get_boxplots_contrasts}{}}}
\subsection{Method \code{get_boxplots_contrasts()}}{
get box with contrast information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$get_boxplots_contrasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-write_boxplots_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-write_boxplots_contrasts}{}}}
\subsection{Method \code{write_boxplots_contrasts()}}{
write boxplots contrasts to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$write_boxplots_contrasts(filename = "boxplots")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename to write to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAnalyse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAnalyse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAnalyse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
