% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEAReportGenerator.R
\name{DEAReportGenerator}
\alias{DEAReportGenerator}
\title{DEAReportGenerator}
\description{
DEAReportGenerator

DEAReportGenerator
}
\details{
Generates all output files for a differential expression analysis.
Uses a DEAnalyse object as data source instead of the legacy GRP2$RES list.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deanalyse}}{DEAnalyse object containing all analysis results}

\item{\code{GRP2}}{ProlfquAppConfig object containing analysis configuration}

\item{\code{fname}}{filename prefix for DEA results}

\item{\code{qcname}}{filename prefix for QC results}

\item{\code{resultdir}}{directory for storing results}

\item{\code{ZIPDIR}}{zip directory path}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAReportGenerator-new}{\code{DEAReportGenerator$new()}}
\item \href{#method-DEAReportGenerator-prep_result_list}{\code{DEAReportGenerator$prep_result_list()}}
\item \href{#method-DEAReportGenerator-write_DEA}{\code{DEAReportGenerator$write_DEA()}}
\item \href{#method-DEAReportGenerator-render_DEA}{\code{DEAReportGenerator$render_DEA()}}
\item \href{#method-DEAReportGenerator-make_boxplots}{\code{DEAReportGenerator$make_boxplots()}}
\item \href{#method-DEAReportGenerator-write_DEA_all}{\code{DEAReportGenerator$write_DEA_all()}}
\item \href{#method-DEAReportGenerator-make_SummarizedExperiment}{\code{DEAReportGenerator$make_SummarizedExperiment()}}
\item \href{#method-DEAReportGenerator-clone}{\code{DEAReportGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialize DEAReportGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$new(deanalyse, GRP2, name = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deanalyse}}{DEAnalyse R6 object with completed analysis}

\item{\code{GRP2}}{ProlfquAppConfig R6 object}

\item{\code{name}}{optional name prefix for output files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-prep_result_list"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-prep_result_list}{}}}
\subsection{Method \code{prep_result_list()}}{
Prepare result list with all analysis outputs for XLSX
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$prep_result_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list containing all analysis results (14 sheets)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-write_DEA"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-write_DEA}{}}}
\subsection{Method \code{write_DEA()}}{
Write DEA results (XLSX, ORA, GSEA files)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$write_DEA(ORA = TRUE, GSEA = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ORA}}{if TRUE write ORA gene lists}

\item{\code{GSEA}}{if TRUE write GSEA rank files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with xlsx_file, ora_files, gsea_files paths
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-render_DEA"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-render_DEA}{}}}
\subsection{Method \code{render_DEA()}}{
Render DEA report using R Markdown
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$render_DEA(
  htmlname,
  markdown = "_Grp2Analysis_V2_R6.Rmd",
  word = FALSE,
  toc = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{htmlname}}{name for the output HTML file}

\item{\code{markdown}}{path to the R Markdown template file}

\item{\code{word}}{logical, if TRUE output Word document, otherwise HTML}

\item{\code{toc}}{logical, if TRUE include table of contents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
path to the output file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-make_boxplots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-make_boxplots}{}}}
\subsection{Method \code{make_boxplots()}}{
Generate boxplots for quality control
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$make_boxplots(boxplot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boxplot}}{logical, if TRUE write boxplots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-write_DEA_all"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-write_DEA_all}{}}}
\subsection{Method \code{write_DEA_all()}}{
Write all DEA results: XLSX, ORA, GSEA, HTML reports, boxplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$write_DEA_all(
  boxplot = TRUE,
  render = TRUE,
  ORA = TRUE,
  GSEA = TRUE,
  markdown = "_Grp2Analysis_V2_R6.Rmd",
  markdown_qc = "_DiffExpQC_R6.Rmd",
  toc = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boxplot}}{if TRUE generate boxplots}

\item{\code{render}}{if TRUE render HTML reports}

\item{\code{ORA}}{if TRUE write ORA gene lists}

\item{\code{GSEA}}{if TRUE write GSEA rank files}

\item{\code{markdown}}{Rmd template for main DEA report}

\item{\code{markdown_qc}}{Rmd template for QC report}

\item{\code{toc}}{if TRUE include table of contents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with dea_file, qc_file, data_files paths
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-make_SummarizedExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-make_SummarizedExperiment}{}}}
\subsection{Method \code{make_SummarizedExperiment()}}{
Create SummarizedExperiment object from analysis results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$make_SummarizedExperiment(
  strip = "~lfq~light",
  .url_builder = prolfquapp::bfabric_url_builder
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strip}}{pattern to strip from rownames}

\item{\code{.url_builder}}{function to build URLs for bfabric}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SummarizedExperiment object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
