% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEAReportGenerator.R
\name{DEAReportGenerator}
\alias{DEAReportGenerator}
\title{DEAReportGenerator}
\description{
DEAReportGenerator

DEAReportGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEAReportGenerator-new}{\code{DEAReportGenerator$new()}}
\item \href{#method-DEAReportGenerator-write_DEA_all}{\code{DEAReportGenerator$write_DEA_all()}}
\item \href{#method-DEAReportGenerator-render_DEA}{\code{DEAReportGenerator$render_DEA()}}
\item \href{#method-DEAReportGenerator-make_boxplots}{\code{DEAReportGenerator$make_boxplots()}}
\item \href{#method-DEAReportGenerator-prep_result_list}{\code{DEAReportGenerator$prep_result_list()}}
\item \href{#method-DEAReportGenerator-make_SummarizedExperiment}{\code{DEAReportGenerator$make_SummarizedExperiment()}}
\item \href{#method-DEAReportGenerator-clone}{\code{DEAReportGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$new(lfqdata, GRP2, prot_annot, Contrasts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-write_DEA_all"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-write_DEA_all}{}}}
\subsection{Method \code{write_DEA_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$write_DEA_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-render_DEA"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-render_DEA}{}}}
\subsection{Method \code{render_DEA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$render_DEA(
  htmlname,
  markdown = "_Grp2Analysis.Rmd",
  word = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-make_boxplots"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-make_boxplots}{}}}
\subsection{Method \code{make_boxplots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$make_boxplots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-prep_result_list"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-prep_result_list}{}}}
\subsection{Method \code{prep_result_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$prep_result_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-make_SummarizedExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-make_SummarizedExperiment}{}}}
\subsection{Method \code{make_SummarizedExperiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$make_SummarizedExperiment(
  strip = "~lfq~light",
  .url_builder = bfabric_url_builder
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEAReportGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEAReportGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEAReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
