% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{diann_read_output}
\alias{diann_read_output}
\title{read DiaNN diann-output.tsv file}
\usage{
diann_read_output(data, Lib.PG.Q.Value = 0.01, PG.Q.Value = 0.05)
}
\arguments{
\item{data}{data frame of DIA-NN report}

\item{Lib.PG.Q.Value}{library protein group q-value threshold}

\item{PG.Q.Value}{protein group q-value threshold}
}
\description{
filter for 2 peptides per protein, and for Q.Value < 0.01 (default)
}
\examples{
\dontrun{
xx <- readr::read_tsv("WU292720_report.tsv")
report2 <- prolfquapp::diann_read_output(xx)
nrow(report2)
}

}
