% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{diann_read_output}
\alias{diann_read_output}
\title{DIANN helper functions}
\usage{
diann_read_output(path, Q.Value = 0.01)
}
\arguments{
\item{Q.Value}{q value threshold}

\item{diann.path}{path to diann-output.tsv}

\item{fasta.file}{path to fasta file}

\item{nrPeptides}{peptide threshold}
}
\description{
DIANN helper functions
}
