% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{extract_contrasts}
\alias{extract_contrasts}
\title{extract contrast from annotation file}
\usage{
extract_contrasts(annot, prefix = "G_", group = "group")
}
\description{
extract contrast from annotation file
}
\examples{

annot <- data.frame(names = c("a1","b1"), group= c("a","b"), ddd = c("T","C"))
testthat::expect_error(extract_contrasts(annot))
annot$control <- annot$ddd
contrast <- extract_contrasts(annot)
stopifnot(contrast == "G_a - G_b")

annot$Contrast <- c("G_a - G_b","G_b - G_a")
annot$ContrastName <- c("a_vs_b","b_vs_a")
annot$control <- NULL
ct <- extract_contrasts(annot)
stopifnot(length(ct) == 2)
}
