% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{feature_annotation_collapse_to_single_row}
\alias{feature_annotation_collapse_to_single_row}
\title{get best feature annotation.}
\usage{
feature_annotation_collapse_to_single_row(x)
}
\description{
get best feature annotation.
}
\examples{
if (FALSE) {
  x <- readr::read_csv("test/out_results_zip/mzmine/result_annotations.csv")
  x
  feature_annotation_collapse_to_single_row(x)
}
}
