% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMzMine.R
\name{feature_annotation_collapse_to_single_row}
\alias{feature_annotation_collapse_to_single_row}
\title{get best feature annotation.}
\usage{
feature_annotation_collapse_to_single_row(x)
}
\description{
get best feature annotation.
}
\examples{
if (FALSE) {
  x <- readr::read_csv("WU323671_mzMine_o35537_WpH9V2_neg_v2_result/mzmine/result_annotations.csv")
  feature_annotation_collapse_to_single_row(x)
}
}
