% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DEA_reports.R
\name{generate_DEA_reports}
\alias{generate_DEA_reports}
\title{Generate differential expression analysis reports}
\usage{
generate_DEA_reports(lfqdata, GRP2, prot_annot)
}
\description{
Generate differential expression analysis reports
}
