% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report.R
\name{generate_DEA_reports2}
\alias{generate_DEA_reports2}
\title{will replace generate_DEA_reports}
\usage{
generate_DEA_reports2(lfqdata, GRP2, prot_annot, Contrasts)
}
\description{
will replace generate_DEA_reports
}
