% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot_from_FASTA.R
\name{get_annot_from_fasta}
\alias{get_annot_from_fasta}
\title{get_annot_from_fasta}
\usage{
get_annot_from_fasta(
  fasta.files,
  pattern_decoys = "^REV_|^rev_",
  isUniprot = TRUE,
  min_length = 7,
  max_length = 30,
  include_seq = FALSE
)
}
\description{
get_annot_from_fasta
}
\examples{

fasta_conn <- textConnection(prolfquapp:::.getSequences())
testthat::expect_error(prolfquapp::get_annot_from_fasta(fasta_conn, pattern_decoys = "" ))
close(fasta_conn)
fasta_conn <- textConnection(prolfquapp:::.getSequences())
prolfquapp::get_annot_from_fasta(fasta_conn, pattern_decoys = "^REV_|^rev" )
close(fasta_conn)

}
