% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{get_config}
\alias{get_config}
\title{get configuration from yaml if exists}
\usage{
get_config(yamlfile, WORKUNITID = "HelloWorld", ORDERID = "123")
}
\description{
get configuration from yaml if exists
}
\examples{


get_config()
}
