% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R, R/tidyMS_MaxQuant.R
\name{get_MQ_peptide_files}
\alias{get_MQ_peptide_files}
\title{get petpide.txt and fasta file location in folder}
\usage{
get_MQ_peptide_files(path)

get_MQ_peptide_files(path)
}
\value{
list with paths to data and fasta

list with paths to data and fasta
}
\description{
get petpide.txt and fasta file location in folder

get petpide.txt and fasta file location in folder
}
