% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Configuration.R
\name{make_DEA_config_R6}
\alias{make_DEA_config_R6}
\title{create GRP2 configuration.
Use this function if there is no Yaml Input.}
\usage{
make_DEA_config_R6(
  ZIPDIR = ".",
  PROJECTID = "",
  ORDERID = "",
  WORKUNITID = "",
  Normalization = c("none", "vsn", "quantile", "robscale"),
  aggregation = c("medpolish", "top3", "lmrob"),
  diff_threshold = 1,
  FDR_threshold = 0.1,
  nr_peptides = 1,
  removeContaminants = FALSE,
  removeDecoys = FALSE,
  patternDecoys = "^REV_",
  patternContaminants = "^zz",
  application = "FragPipeTMT"
)
}
\arguments{
\item{patternDecoys}{default "^REV_"}

\item{patternContaminants}{default "^zz_"}
}
\description{
create GRP2 configuration.
Use this function if there is no Yaml Input.
}
\examples{
DEAconfig <- make_DEA_config_R6()
R6list <- R6_extract_values(DEAconfig)

}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{R6_extract_values}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
