% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R, R/SE_helpers.R,
%   R/make_DEA_report.R
\name{make_DEA_report}
\alias{make_DEA_report}
\alias{render_DEA}
\alias{make_SummarizedExperiment}
\alias{write_DEA}
\title{Create DEA report in html and write data to xlsx table}
\usage{
make_DEA_report(lfqdata, protAnnot, GRP2)

render_DEA(
  GRP2,
  outpath,
  htmlname = "Result2Grp",
  word = FALSE,
  markdown = "_Grp2Analysis.Rmd"
)

make_SummarizedExperiment(
  GRP2,
  colname = NULL,
  rowname = NULL,
  strip = "~lfq~light",
  .url_builder = bfabric_url_builder
)

write_DEA(GRP2, outpath, xlsxname = "AnalysisResults", write = TRUE)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}

\item{outpath}{path to place output}

\item{htmlname}{name for html file}

\item{word}{default FALSE, if true create word document.s}

\item{markdown}{which file to render}

\item{xlsxname}{file name for xlsx}

\item{startdata}{table in long format}

\item{atable}{AnalysisTableAnnotation annotate startdata table}

\item{protein_annot}{column with protein description e.g. (fasta header)}
}
\value{
SummarizedExperiment
}
\description{
For use examples see run_scripts directory
}
\examples{

istar <- prolfqua::sim_lfq_data_peptide_config(Nprot = 100)
data <- istar$data
data$Description <-"AAAAA"
protein_annot <- data |> dplyr::select(protein_Id, description = Description) |> dplyr::distinct()
GRP2 <- list()
GRP2$Bfabric <- list()
GRP2$Bfabric$projectID <- "3765"
GRP2$Bfabric$projectName <- "Order_26863"
GRP2$Bfabric$orderID <- "3765"

GRP2$Bfabric$workunitID <- "2057368.zip"
GRP2$Bfabric$inputID <- "2057368.zip"
GRP2$Bfabric$inputURL <- "https://www.fgcz.ch"

#at least 2 peptides per protein
GRP2$pop <- list()
GRP2$pop$transform <- "vsn"
GRP2$pop$aggregate <- "medpolish"
GRP2$pop$Diffthreshold <- 0.5
GRP2$pop$FDRthreshold <- 0.25
GRP2$pop$Contrasts <- c(b_vs_a = "group_A - group_Ctrl")
GRP2$Software <- "MaxQuant"


lfqdata <- prolfqua::LFQData$new(data, istar$config)
lfqdata$remove_small_intensities()
aggregator <- lfqdata$get_Aggregator()

aggregator$sum_topN()
lfqdata <- aggregator$lfq_agg

GRP2$pop$nrPeptides <- 2

GRP2$pop$revpattern = "^REV_"
GRP2$pop$contpattern = "^zz|^CON__"

GRP2$pop$removeCon = TRUE
GRP2$pop$removeDecoys = TRUE


protAnnot <- prolfqua::ProteinAnnotation$new(
    lfqdata,
    row_annot = protein_annot)

grp <- make_DEA_report(lfqdata, protAnnot, GRP2)
st <- grp$RES$transformedlfqData$get_Stats()
bb <- st$stats()
sr <- grp$RES$transformedlfqData$get_Summariser()
int <- sr$interaction_missing_stats()
res <- write_DEA(grp,".", write=FALSE)
se <- make_SummarizedExperiment(grp)
\dontrun{
render_DEA(grp, ".")
render_DEA(grp, "." ,word = TRUE)

}

}
\concept{workflow}
