% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATE.R
\name{make_DEA_report}
\alias{make_DEA_report}
\alias{render_DEA}
\title{Create DEA report in html and write data to xlsx table}
\usage{
make_DEA_report(lfqdata, protAnnot, GRP2)

render_DEA(
  GRP2,
  outpath,
  htmlname = "Result2Grp",
  word = FALSE,
  markdown = "_Grp2Analysis.Rmd"
)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}

\item{outpath}{path to place output}

\item{htmlname}{name for html file}

\item{word}{default FALSE, if true create word document.s}

\item{markdown}{which file to render}

\item{startdata}{table in long format}

\item{atable}{AnalysisTableAnnotation annotate startdata table}

\item{protein_annot}{column with protein description e.g. (fasta header)}
}
\description{
For use examples see run_scripts directory
}
\seealso{
Other workflow: 
\code{\link{make_SummarizedExperiment}()},
\code{\link{write_DEA}()}

Other workflow: 
\code{\link{make_SummarizedExperiment}()},
\code{\link{write_DEA}()}
}
\concept{workflow}
