% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2GrpAnalysis.R
\name{make_DEA_report}
\alias{make_DEA_report}
\alias{write_DEA}
\alias{render_DEA}
\title{Create DEA report in html and write data to xlsx table}
\usage{
make_DEA_report(lfqdata, prot_annot, GRP2)

write_DEA(GRP2, outpath, xlsxname = "AnalysisResults")

render_DEA(
  GRP2,
  outpath,
  htmlname = "Result2Grp",
  word = FALSE,
  markdown = "_Grp2Analysis.Rmd"
)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}

\item{outpath}{path to place output}

\item{xlsxname}{file name for xlsx}

\item{htmlname}{name for html file}

\item{word}{default FALSE, if true create word document.s}

\item{markdown}{which file to render}

\item{startdata}{table in long format}

\item{atable}{AnalysisTableAnnotation annotate startdata table}

\item{protein_annot}{column with protein description e.g. (fasta header)}
}
\description{
For use examples see run_scripts directory
}
\examples{
library(prolfquapp)
library(prolfqua)
istar <- prolfqua_data('data_ionstar')$filtered()
data <- istar$data |> dplyr::filter(protein_Id \%in\% sample(protein_Id, 100))
data$Description <-"AAAAA"
protein_annot <- data |> dplyr::select(protein_Id, Description) |> dplyr::distinct()
GRP2 <- list()
GRP2$Bfabric <- list()
GRP2$Bfabric$projectID <- "3765"
GRP2$Bfabric$projectName <- "Order_26863"
GRP2$Bfabric$orderID <- "3765"

GRP2$Bfabric$workunitID <- "2057368.zip"
GRP2$Bfabric$inputID <- "2057368.zip"
GRP2$Bfabric$inputURL <- "https://www.fgcz.ch"

#at least 2 peptides per protein
GRP2$pop <- list()
GRP2$pop$transform <- "vsn"
GRP2$pop$aggregate <- "medpolish"
GRP2$pop$Diffthreshold <- 0.5
GRP2$pop$FDRthreshold <- 0.25
GRP2$pop$Contrasts <- c(b_vs_a = "dilution.b - dilution.a")

GRP2$Software <- "MaxQuant"

data <- dplyr::filter(data, dilution. == "a" |  dilution. == "b")
atab <- AnalysisTableAnnotation$new()
atab$ident_qValue = "pep"
atab$fileName = "raw.file"
atab$hierarchy["protein_Id"] = "protein_Id"
atab$hierarchy["peptide_Id"] = "peptide_Id"
atab$factors["dilution."] = "dilution."
atab$set_response("peptide.intensity")
atab$isotopeLabel = "isotope"
config <- prolfqua::AnalysisConfiguration$new(atab)

lfqdata <- prolfqua::LFQData$new(data, config)
lfqdata$remove_small_intensities()
aggregator <- lfqdata$get_Aggregator()

aggregator$sum_topN()
lfqdata <- aggregator$lfq_agg

GRP2$pop$nrPeptides <- 2

GRP2$pop$Contrasts <- c("avsb" = "dilution.a - dilution.b")
GRP2$pop$revpattern = "^REV_"
GRP2$pop$contpattern = "^zz|^CON__"
GRP2$pop$remove = TRUE
grp <- make_DEA_report(lfqdata, protein_annot, GRP2)
st <- grp$RES$transformedlfqData$get_Stats()
bb <- st$stats()
sr <- grp$RES$transformedlfqData$get_Summariser()
int <- sr$interaction_missing_stats()

\dontrun{
render_DEA(grp, ".")
render_DEA(grp, "." ,word = TRUE)
write_DEA(grp,".")
}

}
\concept{workflow}
