% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_vs_percent.R
\name{plot_abundance_vs_percent}
\alias{plot_abundance_vs_percent}
\title{Plot relative protein abundance as a function of rank by abundance}
\usage{
plot_abundance_vs_percent(
  sr,
  top_percent = 5,
  top_N = 10,
  factors = TRUE,
  log = FALSE,
  colors = c(`^REV_` = "red", `^CON_` = "orange"),
  cumulative = TRUE
)
}
\value{
ggplot2
}
\description{
Plot relative protein abundance as a function of rank by abundance
}
\examples{
library(prolfqua)
istar <- prolfqua_data('data_ionstar')
istar$config <- old2new(istar$config)
data <- istar$data |> dplyr::filter(protein_Id \%in\% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
sr <- lfqdata$get_Summariser()

plot_abundance_vs_percent(sr, top_percent = 20, factors = FALSE, cumulative = TRUE)
plot_abundance_vs_percent(sr, top_percent = 3, factors = FALSE, cumulative = FALSE)
plot_abundance_vs_percent(sr, top_percent = 20, factors = TRUE, cumulative = TRUE)
plot_abundance_vs_percent(sr, top_percent = 20, factors = TRUE, cumulative = FALSE)
plot_abundance_vs_percent(sr, top_percent = NULL, factors = TRUE, cumulative = FALSE)
}
