% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_vs_percent.R
\name{plot_abundance_vs_percent}
\alias{plot_abundance_vs_percent}
\title{Plot relative protein abundance as a function of rank by abundance}
\usage{
plot_abundance_vs_percent(
  percInfo,
  cfg_table,
  top_N = 10,
  factors = TRUE,
  colors = c(`^REV_` = "red", `^CON_` = "orange"),
  cumulative = TRUE,
  group = "BB"
)
}
\arguments{
\item{sr}{LFQDataSummarizer}

\item{top_percent}{deprecate}
}
\value{
ggplot2
}
\description{
Plot relative protein abundance as a function of rank by abundance
}
\examples{
library(prolfqua)
istar <- prolfqua_data('data_ionstar')
istar$config <- old2new(istar$config)
data <- istar$data |> dplyr::filter(protein_Id \%in\% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
sr <- lfqdata$get_Summariser()

plot_abundance_vs_percent(sr, top_N = 6, factors = FALSE, cumulative = TRUE)
pd <- plot_abundance_vs_percent(sr, top_N = NULL, factors = FALSE, cumulative = FALSE)

plot_abundance_vs_percent(sr, top_N = 4, factors = TRUE, cumulative = FALSE)
plot_abundance_vs_percent(sr, top_N = NULL, factors = TRUE, cumulative = TRUE)


}
