% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_vs_percent.R
\name{plot_abundance_vs_percent}
\alias{plot_abundance_vs_percent}
\title{Plot relative protein abundance as a function of rank by abundance}
\usage{
plot_abundance_vs_percent(
  percInfo,
  cfg_config,
  top_N = 10,
  factors = TRUE,
  colors = c(`^REV_` = "red", `^CON_` = "orange"),
  columnAb = "abundance_percent",
  group = "BB",
  alpha = 1,
  logY = TRUE
)
}
\arguments{
\item{percInfo}{data frame with percentage abundance info}

\item{cfg_config}{AnalysisConfiguration object}

\item{top_N}{number of top proteins to label}

\item{factors}{if TRUE facet by factors}

\item{colors}{named vector of colors for special proteins}

\item{columnAb}{column name for abundance values}

\item{group}{crosstalk group identifier}

\item{alpha}{point transparency}

\item{logY}{if TRUE use log10 y-axis}
}
\value{
ggplot2
}
\description{
Plot relative protein abundance as a function of rank by abundance
}
\examples{

library(prolfqua)
istar <- prolfqua::prolfqua_data('data_ionstar')
istar$config <- prolfqua::old2new(istar$config)
data <- istar$data |> dplyr::filter(protein_Id \%in\% sample(protein_Id, 100))
lfqdata <- prolfqua::LFQData$new(data, istar$config)
sr <- lfqdata$get_Summariser()
undebug(plot_abundance_vs_percent)
plot_abundance_vs_percent(sr$percentage_abundance(),
 lfqdata$config,
 top_N = 6, factors = FALSE, logY = TRUE)

pd <- plot_abundance_vs_percent(sr$percentage_abundance(),lfqdata$config, top_N = NULL, factors = FALSE)
plot_abundance_vs_percent(sr$percentage_abundance(),lfqdata$config, top_N = 4, factors = TRUE)
plot_abundance_vs_percent(sr$percentage_abundance(),lfqdata$config, top_N = NULL, factors = TRUE)


}
