% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundDisc.R
\name{preprocess_CD}
\alias{preprocess_CD}
\title{load compound discoverer (CD) files}
\usage{
preprocess_CD(in_file, annotation, .func_massage = prolfquapp::massage_CD)
}
\arguments{
\item{in_file}{excel file produced by CD}

\item{annotation}{list returned by `read_annotation` function}
}
\description{
load compound discoverer (CD) files
}
