% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_CompoundDisc.R
\name{preprocess_CD}
\alias{preprocess_CD}
\title{load compound discoverer (CD) files}
\usage{
preprocess_CD(xdl, annotation, .func_massage = prolfquapp::massage_CD)
}
\arguments{
\item{annotation}{list returned by `read_annotation` function}

\item{in_file}{excel file produced by CD}
}
\description{
load compound discoverer (CD) files
}
