% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{preprocess_DIANN}
\alias{preprocess_DIANN}
\title{preprocess DIANN ouput, filter by q_value and nr_peptides}
\usage{
preprocess_DIANN(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev",
  q_value = 0.01,
  hierarchy_depth = 1,
  nr_peptides = 1
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess DIANN ouput, filter by q_value and nr_peptides
}
\examples{
\dontrun{
x <- get_DIANN_files("inst/application/DIANN/2706527/")

annotation <- file.path("inst/application/DIANN/2706527/dataset.csv") |>
  readr::read_csv() |>
  prolfquapp::read_annotation(QC = TRUE)
x$fasta
undebug(preprocess_DIANN)
xd <- preprocess_DIANN(x$data, x$fasta, annotation)
xd$lfqdata$hierarchy_counts()
xd <- preprocess_DIANN(x$data, x$fasta, annotation, nr_peptides = 2)
xd$lfqdata$hierarchy_counts()
}
}
