% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DIANN.R
\name{preprocess_DIANN}
\alias{preprocess_DIANN}
\title{preprocess DIANN ouput, filter by q_value and nr_peptides}
\usage{
preprocess_DIANN(
  quant_data,
  fasta_file,
  annotation,
  q_value = 0.01,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev"
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess DIANN ouput, filter by q_value and nr_peptides
}
\examples{
\dontrun{
x <- get_DIANN_files("inst/application/DIANN/2517219/")

annotation <- file.path("inst/application/DIANN/2517219/dataset.csv") |>
 readr::read_csv() |> prolfquapp::read_annotation(QC = TRUE)
 x$fasta
xd <- preprocess_DIANN(x$data, x$fasta, annotation)
}
}
