% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dummy.R
\name{preprocess_dummy}
\alias{preprocess_dummy}
\title{preprocess_dummy}
\usage{
preprocess_dummy(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev"
)
}
\description{
preprocess_dummy
}
