% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{preprocess_FP_PSM}
\alias{preprocess_FP_PSM}
\title{preprocess FP psm, filter by purity_threshold and PeptideProphetProb}
\usage{
preprocess_FP_PSM(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_",
  purity_threshold = 0.5,
  PeptideProphetProb = 0.9,
  hierarchy_depth = 1,
  parse_fun = tidy_FragPipe_psm
)
}
\arguments{
\item{quant_data}{path to quantification data file(s)}

\item{fasta_file}{path to fasta file(s)}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{purity_threshold}{purity threshold for filtering}

\item{PeptideProphetProb}{PeptideProphet probability threshold}

\item{hierarchy_depth}{hierarchy depth for aggregation}

\item{parse_fun}{function for parsing PSM files}
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess FP psm, filter by purity_threshold and PeptideProphetProb
}
