% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{preprocess_FP_PSM}
\alias{preprocess_FP_PSM}
\title{preprocess FP psm, filter by purity_threshold and PeptideProphetProb}
\usage{
preprocess_FP_PSM(
  quant_data,
  fasta_file,
  annotation,
  purity_threshold = 0.5,
  PeptideProphetProb = 0.9,
  column_before_quants = c("Quan Usage", "Mapped Proteins"),
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess FP psm, filter by purity_threshold and PeptideProphetProb
}
