% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_MaxQuant.R
\name{preprocess_MQ_peptide}
\alias{preprocess_MQ_peptide}
\title{preprocess MQ peptide}
\usage{
preprocess_MQ_peptide(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\description{
preprocess MQ peptide
}
