% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{preprocess_MSstats}
\alias{preprocess_MSstats}
\title{preprocess MSstats file coming from FragPipe}
\usage{
preprocess_MSstats(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_",
  hierarchy_depth = 1
)
}
\arguments{
\item{quant_data}{path to MSstats csv file}

\item{fasta_file}{path to fasta file(s)}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{hierarchy_depth}{hierarchy depth for aggregation}
}
\description{
preprocess MSstats file coming from FragPipe
}
