% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{preprocess_MSstats}
\alias{preprocess_MSstats}
\title{preprocess MQ peptide}
\usage{
preprocess_MSstats(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_",
  hierarchy_depth = 1
)
}
\description{
preprocess MQ peptide
}
