% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{preprocess_MSstats_FPDIA}
\alias{preprocess_MSstats_FPDIA}
\title{preprocess MSstats fragpipe}
\usage{
preprocess_MSstats_FPDIA(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "",
  pattern_decoys = ""
)
}
\description{
preprocess MSstats fragpipe
}
