% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{preprocess_mzMine}
\alias{preprocess_mzMine}
\title{preprocess mzMine input}
\usage{
preprocess_mzMine(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = NULL,
  pattern_decoys = NULL,
  annotated = FALSE
)
}
\description{
preprocess mzMine input
}
