% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MzMine.R
\name{preprocess_mzMine}
\alias{preprocess_mzMine}
\title{preprocess mzMine input}
\usage{
preprocess_mzMine(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = NULL,
  pattern_decoys = NULL,
  annotated = FALSE
)
}
\arguments{
\item{quant_data}{path to mzMine features csv file}

\item{fasta_file}{path to annotations csv file}

\item{annotation}{annotation list from read_annotation}

\item{pattern_contaminants}{regex pattern for contaminants}

\item{pattern_decoys}{regex pattern for decoys}

\item{annotated}{if TRUE only keep annotated features}
}
\description{
preprocess mzMine input
}
\examples{

if(FALSE){
xd <- "outputs-20250407T1707/bfabric/input_dataset.tsv"
annot <- readr::read_tsv(xd)

annotation <- read_annotation(annot, QC = TRUE)
xd <- "outputs-20250407T1707/"
files <- get_mzMine_files(path)
files
undebug(preprocess_mzMine)
res <- preprocess_mzMine(files$data, files$fasta , annotation)
dim(res$lfqdata$data)
res <- preprocess_mzMine(files$data, files$fasta , annotation, annotated = TRUE)
dim(res$lfqdata$data)
}
}
