% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMzMine.R
\name{preprocess_mzMine}
\alias{preprocess_mzMine}
\title{preprocess mzMine input}
\usage{
preprocess_mzMine(quant_data, fasta_file, annotation)
}
\description{
preprocess mzMine input
}
\examples{
xd <- "outputs-20250407T1707/bfabric/input_dataset.tsv"
annot <- readr::read_tsv(xd)
# annot$sample_type <- NULL
annotation <- read_annotation(annot, QC = TRUE)
annotation$annot

res <- preprocess_mzMine("outputs-20250407T1707/mzmine/result_features.csv", NULL , annotation)

}
