% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\name{preprocess_software}
\alias{preprocess_software}
\title{collects preprocess methods for various software}
\usage{
preprocess_software(
  indir,
  annotation,
  preprocess_functions,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^rev_|^REV_",
  extreader = NULL
)
}
\description{
collects preprocess methods for various software
}
\examples{
# example code
annot <- data.frame(
  file = c("a1.raw", "a2.raw", "a3.raw", "a4.raw"),
  name = c("aa", "ba", "aa", "ba"),
  group = c("a", "a", "b", "b")
)

annot <- read_annotation(annot, QC = TRUE)
preprocess_functions <- prolfquapp::prolfqua_preprocess_functions[["DUMMY"]]
res <- preprocess_software(".", annot, preprocess_functions)

xx <- prolfquapp::ExternalReader$new()
xx$extra_args <- "list()"
xx$get_files <- "prolfquapp::get_dummy_files"
xx$preprocess <- "prolfquapp::preprocess_dummy"
res <- preprocess_software(".", annotation = annot, preprocess_functions = xx)
xx <- prolfquapp::ExternalReader$new()

}
