% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\name{preprocess_software}
\alias{preprocess_software}
\title{collects preprocess methods for various software}
\usage{
preprocess_software(
  indir,
  annotation,
  software,
  preprocess_functions_str,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^rev_|^REV_"
)
}
\description{
collects preprocess methods for various software
}
\examples{
# example code
annot <- data.frame(
file = c("a1.raw","a2.raw","a3.raw","a4.raw"),
name = c("aa","ba","aa","ba"),
group = c("a","a","b","b"))
annot <- read_annotation(annot, QC = TRUE)
res <- preprocess_software(".",annot, prolfq_preprocess_functions, software = "DUMMY" )
}
