% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\name{preprocess_software}
\alias{preprocess_software}
\title{collects preprocess methods for various software}
\usage{
preprocess_software(
  indir,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^rev_|^REV_",
  software = c("DIANN", "FP_TMT", "FP_multisite", "MAXQUANT", "MSSTATS")
)
}
\description{
collects preprocess methods for various software
}
