% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Configuration.R
\name{ProcessingOptions}
\alias{ProcessingOptions}
\title{processing options R6 class}
\description{
processing options R6 class

processing options R6 class
}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{R6_extract_values}()},
\code{\link{set_list_to_R6}()}
}
\concept{ProlfquAppConfig}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transform}}{data transformation method}

\item{\code{aggregate}}{protein abundance estimation method}

\item{\code{diff_threshold}}{difference threshold}

\item{\code{FDR_threshold}}{FDR threshold}

\item{\code{remove_cont}}{should contaminants be removed}

\item{\code{remove_decoys}}{should decoys be removed}

\item{\code{pattern_decoys}}{decoy patterns}

\item{\code{pattern_contaminants}}{pattern contaminants}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProcessingOptions-clone}{\code{ProcessingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
