% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_software.R
\docType{data}
\name{prolfqua_preprocess_functions}
\alias{prolfqua_preprocess_functions}
\title{Mapping of software to string representations of functions and arguments
Preprocess Functions Mapping}
\format{
An object of class \code{list} of length 15.
}
\usage{
prolfqua_preprocess_functions
}
\description{
A list containing the mappings of software to string representations of functions and arguments.
This is used to dynamically process different proteomics software outputs.
}
\keyword{datasets}
