% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{read_annotation}
\alias{read_annotation}
\title{read annotation files}
\usage{
read_annotation(dsf, REPEATED = TRUE, SAINT = FALSE)
}
\value{
list with annot (annotation table), atable (analtysis table configuration), contrasts list with contrasts.
}
\description{
read annotation files
}
