% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{read_annotation}
\alias{read_annotation}
\title{read annotation files}
\usage{
read_annotation(dsf, repeated = TRUE, SAINT = FALSE, prefix = "G_", QC = FALSE)
}
\arguments{
\item{dsf}{annotation table}

\item{repeated}{is this a repeated measurement}

\item{SAINT}{is this a SAINTexpress analysis}
}
\value{
list with annot (annotation table), atable (analtysis table configuration), contrasts list with contrasts.
}
\description{
read annotation files
}
\examples{
annot <- data.frame(
file = c("a1.raw","a2.raw","a3.raw","a4.raw"),
name = c("aa","ba","aa","ba"),
group = c("a","a","b","b"))
read_annotation(annot, QC = TRUE)

}
