% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{read_BF_yamlR6}
\alias{read_BF_yamlR6}
\title{read yaml file and convert to R6 configuration object}
\usage{
read_BF_yamlR6(ymlfile, application = "DIANN")
}
\arguments{
\item{ymlfile}{path to yaml configuration file}

\item{application}{software application name}
}
\value{
ProlfquAppConfig R6 object
}
\description{
read yaml file and convert to R6 configuration object
}
\examples{
if (FALSE) {
  yfile <- prolfqua::find_package_file("prolfquapp", "application/DIANN/config.yaml")
  file.exists(yfile)
  config <- read_BF_yamlR6(yfile)
}

}
