% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{read_BF_yamlR6}
\alias{read_BF_yamlR6}
\title{read yaml file}
\usage{
read_BF_yamlR6(ymlfile, application = "DIANN")
}
\value{
list with applications parameters
}
\description{
read yaml file
}
\examples{
if (FALSE) {
  yfile <- prolfqua::find_package_file("prolfquapp", "application/DIANN/config.yaml")
  file.exists(yfile)
  config <- read_BF_yamlR6(yfile)
}

}
