% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSstats.R
\name{read_msstats}
\alias{read_msstats}
\title{read MSstats.csv files and rollup to ProteinSequence level.}
\usage{
read_msstats(file)
}
\arguments{
\item{file}{path to MSstats csv file}
}
\description{
read MSstats.csv files and rollup to ProteinSequence level.
}
