% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AppConfiguration.R
\name{set_list_to_R6}
\alias{set_list_to_R6}
\title{set arguments in list config to r6obj}
\usage{
set_list_to_R6(config_list, r6obj_config)
}
\description{
set arguments in list config to r6obj
}
\seealso{
Other ProlfquAppConfig: 
\code{\link{ExternalReader}},
\code{\link{ProcessingOptions}},
\code{\link{ProjectSpec}},
\code{\link{ProlfquAppConfig}},
\code{\link{R6_extract_values}()},
\code{\link{make_DEA_config_R6}()}
}
\concept{ProlfquAppConfig}
