% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_PSM.R
\name{tidy_FragPipe_psm_V2}
\alias{tidy_FragPipe_psm_V2}
\title{read psm.tsv produced by FragPipe and convert into long format}
\usage{
tidy_FragPipe_psm_V2(
  psm_files,
  purity_threshold = 0.5,
  PeptideProphetProb = 0.9,
  abundance_threshold = 0,
  quan_column_prefix = "^Intensity",
  aggregate = TRUE
)
}
\arguments{
\item{purity_threshold}{purity threshold default = 0.5}

\item{PeptideProphetProb}{default 0.9}

\item{aggregate}{aggregate spectra to psm level}

\item{psm_file}{path to psm.tsv file}

\item{column_before_quants}{describes the last column before the quantitative values (this is not consistent with in different versions of FP, default "Quan Usage"}
}
\description{
read psm.tsv produced by FragPipe and convert into long format
}
