% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMzMine.R
\name{tidy_mzMineFeatures}
\alias{tidy_mzMineFeatures}
\title{convert mzmine features to tidy table}
\usage{
tidy_mzMineFeatures(data)
}
\description{
convert mzmine features to tidy table
}
\examples{
# example code
if(FALSE){
file_path = "outputs-20250407T1707/mzmine/result_features.csv"
raw_df <- read_csv(file_path)
res <- tidy_mzMineFeatures(raw_df)
file_path = "WU323671_mzMine_o35537_WpH9V2_neg_v2_result/mzmine/result_features.csv"
raw_df <- read_csv(file_path)
res <- tidy_mzMineFeatures(raw_df)
head(res)
}
}
