% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnnotationProcessor.R
\name{write_annotation_file}
\alias{write_annotation_file}
\title{write dataset to file in csv, tsv, or xlsx format}
\usage{
write_annotation_file(data, file_path)
}
\description{
write dataset to file in csv, tsv, or xlsx format
}
