% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2GrpAnalysis.R
\name{write_DEA_all}
\alias{write_DEA_all}
\title{Generate differential expression analysis reports}
\usage{
write_DEA_all(grp2, name, ZIPDIR, boxplot = TRUE, render = TRUE)
}
\description{
Writes results of DEA see \code{\link{generate_DEA_reports}}
}
