% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{write_DEA_all}
\alias{write_DEA_all}
\title{Generate differential expression analysis reports}
\usage{
write_DEA_all(
  grp2,
  name = "",
  boxplot = TRUE,
  render = TRUE,
  ORA = TRUE,
  GSEA = TRUE,
  markdown = "_Grp2Analysis.Rmd",
  toc = TRUE
)
}
\arguments{
\item{grp2}{ProlfquAppConfig configuration with results}

\item{name}{optional name prefix for output files}

\item{boxplot}{if TRUE generate boxplots}

\item{render}{if TRUE render HTML reports}

\item{ORA}{if TRUE write ORA gene lists}

\item{GSEA}{if TRUE write GSEA rank files}

\item{markdown}{Rmd template to render}

\item{toc}{if TRUE include table of contents}
}
\description{
Writes results of DEA see \code{\link{generate_DEA_reports}}
}
