% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{write_DEA_all}
\alias{write_DEA_all}
\title{Generate differential expression analysis reports}
\usage{
write_DEA_all(
  grp2,
  name = "",
  boxplot = TRUE,
  render = TRUE,
  ORA = TRUE,
  GSEA = TRUE,
  markdown = "_Grp2Analysis.Rmd"
)
}
\description{
Writes results of DEA see \code{\link{generate_DEA_reports}}
}
