% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DEA_report2.R
\name{write_index_html}
\alias{write_index_html}
\title{write index.html file with links to all relevant files:}
\usage{
write_index_html(file_path_list, result_dir)
}
\description{
write index.html file with links to all relevant files:
}
\examples{
.resdir <- "."
write_index_html(prolfquapp:::.test_links,tempdir())
}
